Organize loading rendering and world command scheduling
Keep loading attempt and failure policy in pump, move drawing behind two private operations, and keep focused tests beside their owners. Remove seven redundant or self-fed loading tests and unused milestone fixtures. Group queue admission, replay draining and scheduled batches in a private world command module without moving frame call sites.
