+++ src/app/loading/pump.rs
@@ -1,37 +1,40 @@
-//! App-level loading screen state and parity primitives.
+//! Loading attempt, asset lease, phase progress and presentation handoff.
 //!
 //! This module sits above simulation and owns loading-screen progress behavior
 //! verified from gamemd.exe. It also owns the request/session boundary used by
 //! app loop. The prepared map, prefix and startup move together; remaining
 //! loading stays synchronous and retains the native progress/presentation order.
+//! Private `render` owns frame drawing; this module retains preparation, cadence,
+//! failure and after-present continuation policy. Request/phase regressions live
+//! in the private `tests` child.
+
+mod render;
+
+#[cfg(test)]
+mod tests;
 
 use crate::app::AppState;
 use crate::app::loading::composition::{
-    LoadingCompositionSnapshot, LoadingParticipantId, LoadingStartAssignment, MmpbRegionRect,
+    LoadingCompositionSnapshot, LoadingParticipantId, LoadingStartAssignment,
     RANDOM_MAP_PREVIEW_FILE, build_loading_composition, build_random_map_loading_composition,
-    loading_base_origin,
 };
 use crate::app::loading::fresh_scenario::FreshScenarioLoadContextDescriptor;
 use crate::app::loading::init::{self, MapLoadInitial, MapLoadResult};
-use crate::app::loading::progress_row::{
-    LoadingProgressRowLayout, LoadingProgressRowSnapshot, layout_standard_skirmish_progress_row,
-};
+use crate::app::loading::progress_row::LoadingProgressRowSnapshot;
 use crate::assets::asset_manager::AssetManager;
 use crate::assets::pal_file::Color;
 use crate::assets::pcx_file::PcxFile;
 use crate::map::preview::DecodedPreview;
 use crate::match_bootstrap::{LoadingStartup, PreparedMatchStartup};
-use crate::render::batch::{BatchRenderer, SpriteInstance};
+use crate::render::batch::BatchRenderer;
 use crate::render::bit_font::BitFont;
-use crate::render::draw_state::DrawState;
 use crate::render::gpu::GpuContext;
 use crate::render::loading_screen_chrome::{
-    LoadingArtVariant, LoadingScreenAtlas, LoadingScreenCompositionAtlasInput, LoadingScreenEntry,
+    LoadingArtVariant, LoadingScreenAtlas, LoadingScreenCompositionAtlasInput,
     LoadingScreenWidth, MmpbMarkerRemap, PreparedLoadingPreviewRgba,
     build_loading_screen_atlas_with_composition,
 };
 use crate::render::shell_surface_present::ShellSurfacePresenter;
-use crate::render::shell_text::{ScissorRect, ShellAlign, ShellTextDraw, TextRect, draw_in_rect};
 use crate::rules::color_scheme::{
     ColorSchemeEntry, hsv_to_rgb, scheme_entry_by_name, scheme_entry_for_priority,
     scheme_hsv_by_entry,
@@ -47,38 +50,11 @@ const STANDARD_SKIRMISH_PROGRESS_MAX: f64 = 100.0;
 const PROGRESS_PERCENT_SCALE: f64 = 0.01;
 const PERCENT_DISPLAY_SCALE: f64 = 100.0;
 const FTOL_EPSILON: f64 = 0.000_001;
-const BACKGROUND_DEPTH: f32 = 0.90;
-const PREVIEW_DEPTH: f32 = 0.80;
-const MARKER_DEPTH: f32 = 0.70;
-const TEXT_BACKING_DEPTH: f32 = 0.60;
-const TEXT_DEPTH: f32 = 0.50;
-const TEXT_BACKING_ALPHA: f32 = 159.0 / 255.0;
-const TEXT_BACKING_PADDING: f32 = 2.0;
-/// Solid backing fill (G3) sits just behind the bar so the bar draws over it.
-const SOLID_FILL_DEPTH: f32 = 0.20;
 /// Static loading colors used only when rules `[Colors]` are unavailable
 /// (missing assets / headless); replaced by `resolve_player_colors` once rules
 /// load.
 const FALLBACK_BACKING_RGB: [f32; 3] = [0.22, 0.22, 0.22];
 const FALLBACK_PROGRESS_RAMP: [Color; 16] = [Color::rgb(180, 180, 180); 16];
-const PROGRESS_DEPTH: f32 = 0.10;
-/// Side icon (G4) draws above the background, at the bar's depth.
-const SIDE_ICON_DEPTH: f32 = 0.10;
-/// Row label follows the bar and country insignia.
-const ROW_LABEL_DEPTH: f32 = 0.05;
-
-/// Effective selected-map standard offline Skirmish milestones after first LS draw.
-///
-/// The theater ramp between 13 and 25 is dynamic in gamemd.exe, so it is modeled
-/// by `theater_ramp_changed_values` rather than hardcoded as every integer.
-pub const STANDARD_SKIRMISH_SELECTED_MAP_MILESTONES_AFTER_FIRST_RENDER: &[u32] = &[
-    3, 8, 12, 25, 30, 31, 35, 45, 50, 55, 58, 60, 63, 65, 67, 68, 69, 70, 72, 74, 76, 78, 82, 86,
-    90, 93, 96, 98, 100,
-];
-
-/// Raw native calls that should not create a redraw in the normal selected-map path.
-pub const STANDARD_SKIRMISH_NONADVANCING_RAW_CALLS: &[u32] = &[6, 58, 60, 25];
-
 #[derive(Debug, Clone, PartialEq)]
 pub struct LoadingProgressState {
     max_value: f64,
@@ -1464,135 +1440,7 @@ fn encode_loading_screen(
     else {
         return Ok(LoadingRenderResult::GenericFallback);
     };
-    let Some(atlas) = native.atlas.as_ref() else {
-        return Err(anyhow::anyhow!(
-            "native Skirmish loading atlas was not available for render"
-        ));
-    };
-    let target = state.renderer.shell_surface_presenter.source_render_view();
-
-    let frame_plan = build_native_loading_frame_plan(
-        &state.renderer.bit_font,
-        atlas,
-        native.composition.as_ref(),
-        &native.progress_row,
-        &native.progress,
-        native.backing_rgb,
-        native.text_rgb,
-        [
-            state.renderer.gpu.config.width,
-            state.renderer.gpu.config.height,
-        ],
-    );
-    let instances = frame_plan.instances;
-    let text_draws = frame_plan.text_draws;
-
-    state.renderer.batch_renderer.update_camera(
-        &state.renderer.gpu,
-        state.renderer.gpu.config.width as f32,
-        state.renderer.gpu.config.height as f32,
-        0.0,
-        0.0,
-        1.0,
-    );
-    let Some((buffer, count)) = state
-        .renderer
-        .batch_renderer
-        .create_instance_buffer(&state.renderer.gpu, &instances)
-    else {
-        return Err(anyhow::anyhow!(
-            "native Skirmish loading instances could not be uploaded"
-        ));
-    };
-    let backing_buffers = text_draws
-        .iter()
-        .map(|draw| {
-            state
-                .renderer
-                .batch_renderer
-                .create_instance_buffer(&state.renderer.gpu, &draw.backing)
-        })
-        .collect::<Vec<_>>();
-    let text_buffers = text_draws
-        .iter()
-        .map(|draw| {
-            state
-                .renderer
-                .batch_renderer
-                .create_instance_buffer(&state.renderer.gpu, &draw.text.instances)
-        })
-        .collect::<Vec<_>>();
-
-    let mut pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-        label: Some("Native Loading Screen"),
-        color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-            view: &target,
-            depth_slice: None,
-            resolve_target: None,
-            ops: wgpu::Operations {
-                load: wgpu::LoadOp::Clear(crate::app::types::CLEAR_COLOR),
-                store: wgpu::StoreOp::Store,
-            },
-        })],
-        depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
-            view: &state.renderer.depth_view,
-            depth_ops: Some(wgpu::Operations {
-                load: wgpu::LoadOp::Clear(1.0),
-                store: wgpu::StoreOp::Store,
-            }),
-            stencil_ops: None,
-        }),
-        timestamp_writes: None,
-        occlusion_query_set: None,
-    });
-    state.renderer.batch_renderer.draw_with_buffer_passthrough(
-        &mut pass,
-        &atlas.texture,
-        &buffer,
-        count,
-    );
-    for ((draw, backing_buffer), text_buffer) in text_draws
-        .iter()
-        .zip(backing_buffers.iter())
-        .zip(text_buffers.iter())
-    {
-        if let Some((buffer, count)) = backing_buffer.as_ref() {
-            state.renderer.batch_renderer.draw_with_buffer_passthrough(
-                &mut pass,
-                &atlas.texture,
-                buffer,
-                *count,
-            );
-        }
-        let Some((buffer, count)) = text_buffer.as_ref() else {
-            continue;
-        };
-        let Some(scissor) = clamp_loading_scissor(
-            draw.text.scissor,
-            state.renderer.gpu.config.width,
-            state.renderer.gpu.config.height,
-        ) else {
-            continue;
-        };
-        pass.set_scissor_rect(scissor.x, scissor.y, scissor.w, scissor.h);
-        state.renderer.batch_renderer.draw_with_buffer_passthrough(
-            &mut pass,
-            state.renderer.bit_font.atlas(),
-            buffer,
-            *count,
-        );
-    }
-    pass.set_scissor_rect(
-        0,
-        0,
-        state.renderer.gpu.config.width,
-        state.renderer.gpu.config.height,
-    );
-    drop(pass);
-    state
-        .renderer
-        .shell_surface_presenter
-        .encode_present(encoder, destination);
+    render::encode_native_loading_frame(&state.renderer, native, encoder, destination)?;
     Ok(LoadingRenderResult::NativeRendered)
 }
 
@@ -1686,449 +1534,6 @@ fn normalize_rgb(rgb: [u8; 3]) -> [f32; 3] {
     ]
 }
 
-struct NativeLoadingTextDraw {
-    backing: Vec<SpriteInstance>,
-    text: ShellTextDraw,
-}
-
-struct NativeLoadingFramePlan {
-    instances: Vec<SpriteInstance>,
-    text_draws: Vec<NativeLoadingTextDraw>,
-}
-
-fn native_loading_row_layout(
-    font: &BitFont,
-    atlas: &LoadingScreenAtlas,
-    progress: &LoadingProgressState,
-    render_size: [u32; 2],
-) -> Option<LoadingProgressRowLayout> {
-    if progress.current_value() == 0.0 {
-        return None;
-    }
-    Some(layout_standard_skirmish_progress_row(
-        render_size,
-        [
-            atlas.progress_frame0.pixel_size[0] as i32,
-            atlas.progress_frame0.pixel_size[1] as i32,
-        ],
-        atlas
-            .side_icon
-            .map(|icon| [icon.pixel_size[0] as i32, icon.pixel_size[1] as i32]),
-        font.cell_height() as i32,
-    ))
-}
-
-fn build_native_loading_text_draws(
-    font: &BitFont,
-    atlas: &LoadingScreenAtlas,
-    composition: Option<&LoadingCompositionSnapshot>,
-    row: &LoadingProgressRowSnapshot,
-    row_layout: Option<&LoadingProgressRowLayout>,
-    text_rgb: [f32; 3],
-    row_rgb: [f32; 3],
-) -> Vec<NativeLoadingTextDraw> {
-    let mut draws = Vec::with_capacity(5);
-    if let Some(composition) = composition {
-        if let Some(text) = composition.text.country_name.as_deref() {
-            draws.push(build_native_loading_text_draw(
-                font,
-                atlas,
-                text,
-                composition.text_rects.country_name,
-                text_rgb,
-                ShellAlign::H_RIGHT,
-                true,
-                TEXT_DEPTH,
-            ));
-        }
-        if let Some(text) = composition.text.special_unit.as_deref() {
-            draws.push(build_native_loading_text_draw(
-                font,
-                atlas,
-                text,
-                composition.text_rects.special_unit,
-                [0.0, 0.0, 0.0],
-                ShellAlign::NONE,
-                false,
-                TEXT_DEPTH,
-            ));
-        }
-        if let Some(text) = composition.text.load_brief.as_deref() {
-            draws.push(build_native_loading_text_draw(
-                font,
-                atlas,
-                text,
-                composition.text_rects.load_brief,
-                text_rgb,
-                ShellAlign::NONE,
-                true,
-                TEXT_DEPTH,
-            ));
-        }
-        if let Some(text) = composition.text.loading.as_deref() {
-            draws.push(build_native_loading_text_draw(
-                font,
-                atlas,
-                text,
-                composition.text_rects.loading,
-                text_rgb,
-                ShellAlign::NONE,
-                true,
-                TEXT_DEPTH,
-            ));
-        }
-    }
-    if let Some(layout) = row_layout
-        && !row.label.is_empty()
-        && layout.label_rect.w > 0
-        && layout.label_rect.h > 0
-    {
-        draws.push(build_native_loading_text_draw(
-            font,
-            atlas,
-            &row.label,
-            layout.label_rect,
-            row_rgb,
-            ShellAlign::NONE,
-            false,
-            ROW_LABEL_DEPTH,
-        ));
-    }
-    draws
-}
-
-#[allow(clippy::too_many_arguments)]
-fn build_native_loading_text_draw(
-    font: &BitFont,
-    atlas: &LoadingScreenAtlas,
-    text: &str,
-    rect: crate::ui::shell::geom::RectPx,
-    color: [f32; 3],
-    align: ShellAlign,
-    with_backing: bool,
-    depth: f32,
-) -> NativeLoadingTextDraw {
-    let width = rect.w.max(0) as u32;
-    let height = rect.h.max(0) as u32;
-    let text_rect = TextRect {
-        x: rect.x,
-        y: rect.y,
-        w: width,
-        h: height,
-    };
-    let text_draw = draw_in_rect(font, text, text_rect, color, align, [0.0, 0.0], depth, None);
-    let mut backing = Vec::new();
-    if with_backing && !text_draw.instances.is_empty() {
-        let layout = font.wrap_layout(text, width);
-        let aligned_x = if align.contains(ShellAlign::H_RIGHT) && layout.width < width {
-            rect.x + (width - layout.width) as i32
-        } else if align.contains(ShellAlign::H_CENTER) && layout.width < width {
-            rect.x + ((width - layout.width) / 2) as i32
-        } else {
-            rect.x
-        };
-        push_entry_tinted(
-            &mut backing,
-            atlas.solid_texel,
-            [
-                aligned_x as f32 - TEXT_BACKING_PADDING,
-                rect.y as f32 - TEXT_BACKING_PADDING,
-            ],
-            [
-                layout.width as f32 + TEXT_BACKING_PADDING * 2.0,
-                layout.height.min(height) as f32 + TEXT_BACKING_PADDING * 2.0,
-            ],
-            TEXT_BACKING_DEPTH,
-            [0.0, 0.0, 0.0],
-        );
-        if let Some(instance) = backing.last_mut() {
-            instance.alpha = TEXT_BACKING_ALPHA;
-        }
-    }
-    NativeLoadingTextDraw {
-        backing,
-        text: text_draw,
-    }
-}
-
-fn clamp_loading_scissor(
-    scissor: ScissorRect,
-    render_width: u32,
-    render_height: u32,
-) -> Option<ScissorRect> {
-    let x = scissor.x.min(render_width);
-    let y = scissor.y.min(render_height);
-    let w = scissor.w.min(render_width.saturating_sub(x));
-    let h = scissor.h.min(render_height.saturating_sub(y));
-    (w > 0 && h > 0).then_some(ScissorRect { x, y, w, h })
-}
-
-/// Build the full native loading-screen instance list (background, solid backing
-/// fill, clipped progress bar, side icon) shared by the per-frame render path and
-/// the synchronous-repaint sink.
-fn build_native_loading_instances(
-    atlas: &LoadingScreenAtlas,
-    composition: Option<&LoadingCompositionSnapshot>,
-    progress: &LoadingProgressState,
-    backing_rgb: [f32; 3],
-    row_layout: Option<&LoadingProgressRowLayout>,
-    base_origin: [i32; 2],
-) -> Vec<SpriteInstance> {
-    let mut instances = Vec::with_capacity(12);
-    // The art hangs off the same base origin as the progress row and the text
-    // layers, so an oversized window centers all three together.
-    push_entry(
-        &mut instances,
-        atlas.background,
-        [base_origin[0] as f32, base_origin[1] as f32],
-        BACKGROUND_DEPTH,
-    );
-
-    if let Some(composition) = composition {
-        if let (Some(prepared), Some(preview_entry)) = (composition.preview.as_ref(), atlas.preview)
-        {
-            // gamemd blits the source preview into the fitted destination rect,
-            // resampling the whole image. The aspect fit has already chosen a
-            // destination that preserves the source ratio, so both axes scale;
-            // clipping either one here would cut the map's edge off.
-            push_entry_scaled(
-                &mut instances,
-                preview_entry,
-                [
-                    (prepared.region.x + prepared.fit.pad_x) as f32,
-                    (prepared.region.y + prepared.fit.pad_y) as f32,
-                ],
-                [prepared.fit.width as f32, prepared.fit.height as f32],
-                PREVIEW_DEPTH,
-                [1.0; 3],
-            );
-        }
-        // Markers only exist alongside a preview, and they are cropped to that
-        // preview's region for the same reason gamemd composes them into a
-        // region-sized surface before blitting it.
-        if let Some(prepared) = composition.preview.as_ref() {
-            for marker in &composition.markers {
-                let color_key = scheme_entry_for_priority(i32::from(marker.color_priority)) as u8;
-                let Some(entry) = atlas.mmpb_markers.get(&color_key).copied() else {
-                    continue;
-                };
-                push_entry_clipped(
-                    &mut instances,
-                    entry,
-                    [marker.anchor.screen_x, marker.anchor.screen_y],
-                    MARKER_DEPTH,
-                    prepared.region,
-                );
-            }
-        }
-    }
-
-    // The LS renderer's compose-only state owns no progress row. The selected-map
-    // path advances to 3 before the first confirmed display blit.
-    if progress.current_value() == 0.0 {
-        return instances;
-    }
-
-    let Some(row_layout) = row_layout else {
-        return instances;
-    };
-    let bar_w = atlas.progress_frame0.pixel_size[0];
-    let bar_h = atlas.progress_frame0.pixel_size[1];
-    let bar_origin = [
-        row_layout.bar_origin[0] as f32,
-        row_layout.bar_origin[1] as f32,
-    ];
-
-    // G3: solid backing fill — full bar frame rect (W x H), filled with the
-    // player scheme's `[Colors]` HSV→RGB color, drawn BEFORE the clipped bar so
-    // the bar covers it.
-    push_entry_tinted(
-        &mut instances,
-        atlas.solid_texel,
-        bar_origin,
-        [bar_w, bar_h],
-        SOLID_FILL_DEPTH,
-        backing_rgb,
-    );
-
-    // G2: clipped progress span. The session atlas already contains the
-    // player's 16-shade remap, so preserve its per-pixel colors.
-    let progress_width = progress.fill_width_gamemd_ftol_positive_domain(bar_w as u32);
-    if progress_width > 0 {
-        push_progress_fill(
-            &mut instances,
-            atlas.progress_frame0,
-            bar_origin,
-            progress_width as f32,
-            PROGRESS_DEPTH,
-        );
-    }
-
-    // Country insignia follows the progress span. The atlas has already applied
-    // the verified RGB-magenta key.
-    if let (Some(icon), Some(icon_origin)) = (atlas.side_icon, row_layout.icon_origin) {
-        push_entry(
-            &mut instances,
-            icon,
-            [icon_origin[0] as f32, icon_origin[1] as f32],
-            SIDE_ICON_DEPTH,
-        );
-    }
-
-    instances
-}
-
-#[allow(clippy::too_many_arguments)]
-fn build_native_loading_frame_plan(
-    font: &BitFont,
-    atlas: &LoadingScreenAtlas,
-    composition: Option<&LoadingCompositionSnapshot>,
-    progress_row: &LoadingProgressRowSnapshot,
-    progress: &LoadingProgressState,
-    backing_rgb: [f32; 3],
-    text_rgb: [f32; 3],
-    render_size: [u32; 2],
-) -> NativeLoadingFramePlan {
-    let row_layout = native_loading_row_layout(font, atlas, progress, render_size);
-    let instances = build_native_loading_instances(
-        atlas,
-        composition,
-        progress,
-        backing_rgb,
-        row_layout.as_ref(),
-        loading_base_origin(render_size),
-    );
-    let text_draws = build_native_loading_text_draws(
-        font,
-        atlas,
-        composition,
-        progress_row,
-        row_layout.as_ref(),
-        text_rgb,
-        backing_rgb,
-    );
-    NativeLoadingFramePlan {
-        instances,
-        text_draws,
-    }
-}
-
-/// Acquire a surface frame, render the native loading screen, and present it.
-///
-/// Used by the synchronous-repaint sink to mirror gamemd's per-milestone
-/// `WM_PAINT`. All wgpu ops take `&self`, so only shared references are needed.
-/// Returns an error on acquire/upload failure; the caller treats it as non-fatal.
-fn present_native_loading(
-    gpu: &GpuContext,
-    presenter: &ShellSurfacePresenter,
-    depth_view: &wgpu::TextureView,
-    batch: &BatchRenderer,
-    font: &BitFont,
-    atlas: &LoadingScreenAtlas,
-    composition: Option<&LoadingCompositionSnapshot>,
-    progress_row: &LoadingProgressRowSnapshot,
-    progress: &LoadingProgressState,
-    backing_rgb: [f32; 3],
-    text_rgb: [f32; 3],
-    render_size: [u32; 2],
-) -> anyhow::Result<()> {
-    let output = gpu
-        .surface
-        .get_current_texture()
-        .map_err(|e| anyhow::anyhow!("loading repaint surface texture: {e}"))?;
-    let view = presenter.source_render_view();
-    let mut encoder = gpu
-        .device
-        .create_command_encoder(&wgpu::CommandEncoderDescriptor {
-            label: Some("Native Loading Repaint"),
-        });
-
-    let frame_plan = build_native_loading_frame_plan(
-        font,
-        atlas,
-        composition,
-        progress_row,
-        progress,
-        backing_rgb,
-        text_rgb,
-        render_size,
-    );
-    let instances = frame_plan.instances;
-    let text_draws = frame_plan.text_draws;
-    batch.update_camera(
-        gpu,
-        gpu.config.width as f32,
-        gpu.config.height as f32,
-        0.0,
-        0.0,
-        1.0,
-    );
-    let Some((buffer, count)) = batch.create_instance_buffer(gpu, &instances) else {
-        return Err(anyhow::anyhow!(
-            "loading repaint instances could not be uploaded"
-        ));
-    };
-    let backing_buffers = text_draws
-        .iter()
-        .map(|draw| batch.create_instance_buffer(gpu, &draw.backing))
-        .collect::<Vec<_>>();
-    let text_buffers = text_draws
-        .iter()
-        .map(|draw| batch.create_instance_buffer(gpu, &draw.text.instances))
-        .collect::<Vec<_>>();
-
-    {
-        let mut pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-            label: Some("Native Loading Repaint"),
-            color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                view: &view,
-                depth_slice: None,
-                resolve_target: None,
-                ops: wgpu::Operations {
-                    load: wgpu::LoadOp::Clear(crate::app::types::CLEAR_COLOR),
-                    store: wgpu::StoreOp::Store,
-                },
-            })],
-            depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
-                view: depth_view,
-                depth_ops: Some(wgpu::Operations {
-                    load: wgpu::LoadOp::Clear(1.0),
-                    store: wgpu::StoreOp::Store,
-                }),
-                stencil_ops: None,
-            }),
-            timestamp_writes: None,
-            occlusion_query_set: None,
-        });
-        batch.draw_with_buffer_passthrough(&mut pass, &atlas.texture, &buffer, count);
-        for ((draw, backing_buffer), text_buffer) in text_draws
-            .iter()
-            .zip(backing_buffers.iter())
-            .zip(text_buffers.iter())
-        {
-            if let Some((buffer, count)) = backing_buffer.as_ref() {
-                batch.draw_with_buffer_passthrough(&mut pass, &atlas.texture, buffer, *count);
-            }
-            let Some((buffer, count)) = text_buffer.as_ref() else {
-                continue;
-            };
-            let Some(scissor) =
-                clamp_loading_scissor(draw.text.scissor, gpu.config.width, gpu.config.height)
-            else {
-                continue;
-            };
-            pass.set_scissor_rect(scissor.x, scissor.y, scissor.w, scissor.h);
-            batch.draw_with_buffer_passthrough(&mut pass, font.atlas(), buffer, *count);
-        }
-        pass.set_scissor_rect(0, 0, gpu.config.width, gpu.config.height);
-    }
-
-    presenter.encode_present(&mut encoder, &output.texture);
-    gpu.queue.submit(std::iter::once(encoder.finish()));
-    output.present();
-    Ok(())
-}
-
 fn advance_and_present_native_progress(
     gpu: &GpuContext,
     presenter: &ShellSurfacePresenter,
@@ -2147,7 +1552,7 @@ fn advance_and_present_native_progress(
         return;
     };
     log::debug!(target: "vera20k::loading_attempt", "progress_present raw={raw_percent} effective={effective_percent}");
-    if let Err(err) = present_native_loading(
+    if let Err(err) = render::present_native_loading(
         gpu,
         presenter,
         depth_view,
@@ -2193,7 +1598,7 @@ impl LoadingProgressSink for RenderingProgressSink<'_> {
         let effective_percent = self.cadence.effective_percent(raw_percent);
         if self.progress.advance_progress(effective_percent) {
             log::debug!(target: "vera20k::loading_attempt", "progress_present raw={raw_percent} effective={effective_percent}");
-            if let Err(err) = present_native_loading(
+            if let Err(err) = render::present_native_loading(
                 self.gpu,
                 self.presenter,
                 self.depth_view,
@@ -2216,124 +1621,6 @@ impl LoadingProgressSink for RenderingProgressSink<'_> {
     }
 }
 
-fn push_entry(
-    out: &mut Vec<SpriteInstance>,
-    entry: LoadingScreenEntry,
-    position: [f32; 2],
-    depth: f32,
-) {
-    push_entry_scaled(out, entry, position, entry.pixel_size, depth, [1.0; 3]);
-}
-
-/// Push a quad that resamples the whole source into `size`.
-///
-/// This is the ordinary scaling blit: the full atlas slot is sampled across the
-/// destination rect, so a destination smaller than the source squashes the image
-/// instead of cutting pieces off it.
-fn push_entry_scaled(
-    out: &mut Vec<SpriteInstance>,
-    entry: LoadingScreenEntry,
-    position: [f32; 2],
-    size: [f32; 2],
-    depth: f32,
-    tint: [f32; 3],
-) {
-    out.push(SpriteInstance {
-        position,
-        size,
-        uv_origin: entry.uv_origin,
-        uv_size: entry.uv_size,
-        depth,
-        tint,
-        alpha: 1.0,
-        draw_state: DrawState::default(),
-    });
-}
-
-/// Push a quad cropped to the preview region, dropping it when nothing is left.
-///
-/// gamemd composes the start markers into a surface exactly the size of the
-/// preview region and blits that surface, so a marker whose nudge pushes it past
-/// an edge is cut off there instead of spilling onto the loading art.
-fn push_entry_clipped(
-    out: &mut Vec<SpriteInstance>,
-    entry: LoadingScreenEntry,
-    position: [i32; 2],
-    depth: f32,
-    clip: MmpbRegionRect,
-) {
-    let width = entry.pixel_size[0] as i32;
-    let height = entry.pixel_size[1] as i32;
-    if width <= 0 || height <= 0 {
-        return;
-    }
-    let left = position[0].max(clip.x);
-    let top = position[1].max(clip.y);
-    let right = (position[0] + width).min(clip.x + clip.width);
-    let bottom = (position[1] + height).min(clip.y + clip.height);
-    if right <= left || bottom <= top {
-        return;
-    }
-
-    let visible = [(right - left) as f32, (bottom - top) as f32];
-    let cropped = [(left - position[0]) as f32, (top - position[1]) as f32];
-    out.push(SpriteInstance {
-        position: [left as f32, top as f32],
-        size: visible,
-        uv_origin: [
-            entry.uv_origin[0] + entry.uv_size[0] * cropped[0] / entry.pixel_size[0],
-            entry.uv_origin[1] + entry.uv_size[1] * cropped[1] / entry.pixel_size[1],
-        ],
-        uv_size: [
-            entry.uv_size[0] * visible[0] / entry.pixel_size[0],
-            entry.uv_size[1] * visible[1] / entry.pixel_size[1],
-        ],
-        depth,
-        tint: [1.0, 1.0, 1.0],
-        alpha: 1.0,
-        draw_state: DrawState::default(),
-    });
-}
-
-fn push_entry_tinted(
-    out: &mut Vec<SpriteInstance>,
-    entry: LoadingScreenEntry,
-    position: [f32; 2],
-    size: [f32; 2],
-    depth: f32,
-    tint: [f32; 3],
-) {
-    push_entry_scaled(out, entry, position, size, depth, tint);
-}
-
-/// Push the progress bar's filled span: `PROGBARM.SHP` frame 0 revealed from the
-/// left, full height.
-///
-/// This is the one loading-screen layer that is *clipped* rather than scaled —
-/// the bar sweeps by uncovering more of the same frame, so the U axis is cut at
-/// the fill width while the V axis stays whole. Every other layer scales.
-fn push_progress_fill(
-    out: &mut Vec<SpriteInstance>,
-    entry: LoadingScreenEntry,
-    position: [f32; 2],
-    fill_width: f32,
-    depth: f32,
-) {
-    out.push(SpriteInstance {
-        position,
-        size: [fill_width, entry.pixel_size[1]],
-        uv_origin: entry.uv_origin,
-        uv_size: [
-            entry.uv_size[0] * (fill_width / entry.pixel_size[0]).clamp(0.0, 1.0),
-            entry.uv_size[1],
-        ],
-        depth,
-        tint: [1.0, 1.0, 1.0],
-        alpha: 1.0,
-        draw_state: DrawState::default(),
-    });
-}
-
 fn gamemd_ftol_positive_domain(value: f64) -> i32 {
     debug_assert!(value >= 0.0);
     let nearest = value.round();
@@ -2343,1998 +1630,4 @@ fn gamemd_ftol_positive_domain(value: f64) -> i32 {
 
     // Exact fractional x87 control-word behavior remains a narrow follow-up.
     value as i32
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::skirmish_launch::{
-        AiDifficulty, LaunchCountry, LaunchStartPosition, LaunchTeam, SkirmishAiSlot,
-        SkirmishLaunchMode, SkirmishLaunchOptions, SkirmishLocalSlot,
-    };
-
-    /// Test sink that applies the monotonic gate and records every milestone
-    /// that actually advanced the bar, so the emit sequence can be asserted.
-    struct RecordingProgressSink {
-        progress: LoadingProgressState,
-        emitted: Vec<u32>,
-    }
-
-    impl RecordingProgressSink {
-        fn standard() -> Self {
-            Self {
-                progress: LoadingProgressState::standard_skirmish(),
-                emitted: Vec::new(),
-            }
-        }
-    }
-
-    impl LoadingProgressSink for RecordingProgressSink {
-        fn milestone(&mut self, percent: u32) {
-            if self.progress.advance_progress(percent) {
-                self.emitted.push(percent);
-            }
-        }
-    }
-
-    pub(super) fn test_launch_session(country: LaunchCountry) -> SkirmishLaunchSession {
-        SkirmishLaunchSession {
-            mode: SkirmishLaunchMode {
-                id: 1,
-                ui_name_key: "GUI:Battle".to_string(),
-                tooltip_key: "STT:ModeBattle".to_string(),
-                override_file: "MPBattleMD.ini".to_string(),
-                map_filter: "standard".to_string(),
-                random_maps_allowed: true,
-                allies_allowed: true,
-                must_ally: false,
-            },
-            selected_map_file: Some("mp01t4.map".to_string()),
-            player_name: "Player".to_string(),
-            local: SkirmishLocalSlot {
-                country,
-                country_random: false,
-                color_index: 0,
-                color_random: false,
-                start_position: LaunchStartPosition::Position(0),
-                team: LaunchTeam::None,
-            },
-            opponents: vec![SkirmishAiSlot {
-                country: LaunchCountry::Russia,
-                country_random: false,
-                color_index: 1,
-                color_random: false,
-                start_position: LaunchStartPosition::Position(1),
-                team: LaunchTeam::None,
-                difficulty: AiDifficulty::Easy,
-            }],
-            pre_fill_house_roster:
-                crate::skirmish_launch::PreFillHouseRoster::from_compact_skirmish(1),
-            options: SkirmishLaunchOptions::default(),
-        }
-    }
-
-    struct TestClock(u32);
-
-    impl crate::match_bootstrap::MatchSeedClock for TestClock {
-        fn low_u32(&mut self) -> u32 {
-            self.0
-        }
-
-        fn source(&self) -> crate::match_bootstrap::MatchSeedSource {
-            crate::match_bootstrap::MatchSeedSource::Controlled
-        }
-
-        fn seed_authority_certifying(&self) -> bool {
-            true
-        }
-    }
-
-    fn prepared_startup(next: &mut u64, seed: u32) -> crate::match_bootstrap::PreparedMatchStartup {
-        let launch = test_launch_session(LaunchCountry::America);
-        let accepted = match crate::match_bootstrap::classify_startup_session(&launch) {
-            crate::match_bootstrap::StartupSessionClassification::AcceptedExplicitFixedBattle(
-                accepted,
-            ) => accepted,
-            other => panic!("fixture was not accepted: {other:?}"),
-        };
-        let correlation = crate::match_bootstrap::allocate_match_correlation(next).unwrap();
-        crate::match_bootstrap::prepare_match_startup(correlation, accepted, &mut TestClock(seed))
-    }
-
-    pub(super) fn unverified_seed(value: u32) -> crate::match_bootstrap::MatchSeed {
-        crate::match_bootstrap::MatchSeed {
-            value,
-            source: crate::match_bootstrap::MatchSeedSource::Controlled,
-            seed_authority_certifying: false,
-        }
-    }
-
-    fn prefix_test_map(starts: &[(u8, u16, u16)]) -> crate::map::map_file::MapFile {
-        let mut map =
-            crate::map::rmg::emit::empty_map_file(&crate::map::rmg::RmgOptions::default(), 32, 32);
-        map.waypoints.extend(starts.iter().map(|&(slot, rx, ry)| {
-            (
-                u32::from(slot),
-                crate::map::waypoints::Waypoint {
-                    index: u32::from(slot),
-                    rx,
-                    ry,
-                },
-            )
-        }));
-        map
-    }
-
-    fn generated_preview_with_starts(
-        seed: u16,
-        starts: &[(u8, u16, u16)],
-    ) -> crate::map::rmg::GeneratedMap {
-        crate::map::rmg::GeneratedMap {
-            map_file: prefix_test_map(starts),
-            mapgen_continuation: crate::map::rmg::RmgRng::new(seed).into_continuation(),
-            construction_trace: crate::map::rmg::RmgConstructionTrace::default(),
-            start_waypoints: starts.to_vec(),
-            stages_run: Vec::new(),
-            unfilled_start_slots: 0,
-        }
-    }
-
-    fn accepted_random_map_with_starts(
-        selected_map_file: &str,
-        seed: u16,
-        preview_waypoints: &[(u8, u16, u16)],
-        staged_starts: &[(u8, u16, u16)],
-    ) -> crate::app::shell_random_map::AcceptedRandomMapLaunch {
-        let mut retention = crate::app::shell_random_map::RandomMapGenerationRetention::default();
-        let mut generated = generated_preview_with_starts(seed, staged_starts);
-        generated.map_file = prefix_test_map(preview_waypoints);
-        retention.finish_generation(generated);
-        retention.accept_setup(selected_map_file);
-        retention
-            .take_acceptance_for_loading(Some(selected_map_file))
-            .expect("matching accepted random-map fixture")
-    }
-
-    fn receipt_for(
-        startup: &crate::match_bootstrap::PreparedMatchStartup,
-    ) -> crate::match_bootstrap::RustL0Receipt {
-        let simulation = crate::sim::world::Simulation::with_seed(u64::from(startup.seed.value));
-        crate::match_bootstrap::RustL0Observation {
-            startup,
-            simulation: &simulation,
-            active_correlation: startup.correlation,
-            prior_receipt: None,
-            screen_is_loading: true,
-            spawn_pick_active: false,
-        }
-        .acknowledge()
-        .expect("valid test startup must acknowledge")
-    }
-
-    fn test_audio() -> crate::app::audio_runtime::AppAudioRuntime {
-        crate::app::audio_runtime::AppAudioRuntime {
-            theme: crate::audio::theme::ThemeRuntime::default(),
-            last_theme_poll_ms: None,
-            music_player: None,
-            sfx_player: None,
-            sound_registry: Default::default(),
-            audio_indices: Vec::new(),
-            audio_indices_enabled: false,
-            launcher_audio_available: true,
-            theme_startup_suppressed: false,
-            eva_registry: Default::default(),
-        }
-    }
-
-    /// `begin_loading`'s shell -> scenario boundary: the LOADING request
-    /// (`Start_Scenario @ 0x00683AB0`, `Play_Song(From_Name("LOADING"))` @
-    /// `0x00683D1A`) must resolve the process asset manager *through the
-    /// loading-job lease*, and the audio pump's Theme poll must keep reaching
-    /// AI while that lease is out. Exercises the production leaf that
-    /// `begin_loading` calls and the resolver `pump_audio_service` uses.
-    #[test]
-    fn begin_loading_plays_loading_theme_and_polls_theme_through_the_lease() {
-        let dir =
-            std::env::temp_dir().join(format!("vera20k-loading-theme-{}", std::process::id()));
-        std::fs::create_dir_all(&dir).expect("test asset dir");
-        std::fs::write(
-            dir.join("thememd.ini"),
-            b"[Themes]\n1=INTRO\n2=LOADING\n\n[INTRO]\nName=Intro\nSound=intro\nRepeat=yes\n\n\
-              [LOADING]\nName=Loading\nSound=loading\nRepeat=yes\n",
-        )
-        .expect("write thememd.ini");
-        let mut process_assets = crate::app::process_assets::ProcessAssets::from_startup(
-            Some(AssetManager::from_loose_root_for_test(&dir)),
-            None,
-            None,
-            None,
-        );
-        let mut audio = test_audio();
-        let session = LoadingSession::from_request(LoadingRequest::unverified_legacy_skirmish(
-            test_launch_session(LaunchCountry::America),
-            unverified_seed(1),
-            SkirmishSettings::default(),
-        ));
-
-        let mut slot = None;
-        let mut startup = crate::app::match_runtime::startup::MatchStartup::default();
-        replace_loading_attempt(
-            &mut slot,
-            &mut startup,
-            &mut process_assets,
-            &mut audio,
-            session,
-            500,
-        );
-        let session = slot.as_ref().unwrap();
-
-        assert!(
-            process_assets.manager().is_none() && process_assets.is_leased(),
-            "the manager is leased into the loading job"
-        );
-        let loading = audio.theme.from_name("LOADING");
-        assert!(loading >= 0, "catalog resolves LOADING");
-        let slots = audio.theme.slots();
-        assert_eq!(
-            slots.retained, loading,
-            "Play_Song(LOADING) was issued while the resident slot was empty"
-        );
-        assert_eq!(
-            slots.pending, loading,
-            "LOADING repeats under the loading screen"
-        );
-
-        // The audio pump's poll resolves the leased manager and advances AI.
-        assert!(
-            audio_service_asset_manager(&process_assets, None).is_none(),
-            "resident slot alone cannot serve the poll during a lease"
-        );
-        let assets = audio_service_asset_manager(&process_assets, Some(session))
-            .expect("leased manager serves the Theme poll");
-        audio.update_theme(assets, 600);
-        assert_eq!(audio.last_theme_poll_ms, Some(600));
-    }
-
-    #[test]
-    fn loading_replacement_and_terminal_retirement_preserve_cache_and_admission_order() {
-        let dir =
-            std::env::temp_dir().join(format!("vera20k-loading-owner-{}", std::process::id()));
-        std::fs::create_dir_all(&dir).unwrap();
-        std::fs::write(dir.join("sentinel.bin"), b"first winner").unwrap();
-        let mut assets = crate::app::process_assets::ProcessAssets::from_startup(
-            Some(AssetManager::from_loose_root_for_test(&dir)),
-            None,
-            None,
-            None,
-        );
-        let original = assets
-            .manager()
-            .unwrap()
-            .load_file_from_mix("sentinel.bin")
-            .unwrap();
-        let mut audio = test_audio();
-        let mut startup = crate::app::match_runtime::startup::MatchStartup::default();
-        let mut slot = None;
-        let mut next = 1;
-        let first = prepared_startup(&mut next, 7);
-        let replacement = prepared_startup(&mut next, 8);
-        for prepared in [&first, &replacement] {
-            replace_loading_attempt(
-                &mut slot,
-                &mut startup,
-                &mut assets,
-                &mut audio,
-                LoadingSession::from_request(LoadingRequest::accepted_skirmish(
-                    prepared.clone(),
-                    SkirmishSettings::default(),
-                )),
-                500,
-            );
-            let cached = loading_asset_manager(slot.as_ref().unwrap())
-                .unwrap()
-                .load_file_from_mix("SENTINEL.BIN")
-                .unwrap();
-            assert!(std::sync::Arc::ptr_eq(&original.bytes, &cached.bytes));
-        }
-        assert!(
-            startup
-                .acknowledge(
-                    first.clone(),
-                    Some(&crate::sim::world::Simulation::with_seed(7)),
-                    true,
-                    false
-                )
-                .is_err()
-        );
-        // The installer's resource-only cleanup must preserve the replacement
-        // admission until its actual L0 observation commits.
-        retire_loading_attempt(&mut slot, &mut assets);
-        assert!(slot.is_none() && assets.is_available() && !assets.is_leased());
-        startup
-            .acknowledge(
-                replacement.clone(),
-                Some(&crate::sim::world::Simulation::with_seed(8)),
-                true,
-                false,
-            )
-            .unwrap();
-        assert_eq!(startup.startup(), Some(&replacement));
-
-        for consumed_during_preparation in [false, true] {
-            let attempt = prepared_startup(&mut next, 9);
-            replace_loading_attempt(
-                &mut slot,
-                &mut startup,
-                &mut assets,
-                &mut audio,
-                LoadingSession::from_request(LoadingRequest::accepted_skirmish(
-                    attempt.clone(),
-                    SkirmishSettings::default(),
-                )),
-                600,
-            );
-            if consumed_during_preparation {
-                // Exact production preparation failure before initial selection;
-                // no config exists, while the prior lease is already owned.
-                let err = prepare_loading_session(&mut assets, slot.take().unwrap(), true, None)
-                    .err()
-                    .unwrap();
-                assert!(err.to_string().contains("missing game config"));
-                assert!(assets.is_available());
-            }
-            assert_eq!(
-                retire_failed_loading_attempt(
-                    &mut slot,
-                    &mut startup,
-                    &mut assets,
-                    LoadingFailurePolicy::ReportNativeFailure
-                ),
-                LoadingFailurePolicy::ReportNativeFailure
-            );
-            assert!(slot.is_none() && assets.is_available() && !assets.is_leased());
-            assert!(
-                startup
-                    .acknowledge(
-                        attempt,
-                        Some(&crate::sim::world::Simulation::with_seed(9)),
-                        true,
-                        false
-                    )
-                    .is_err()
-            );
-            let cached = assets
-                .manager()
-                .unwrap()
-                .load_file_from_mix("sentinel.bin")
-                .unwrap();
-            assert!(std::sync::Arc::ptr_eq(&original.bytes, &cached.bytes));
-        }
-
-        replace_loading_attempt(
-            &mut slot,
-            &mut startup,
-            &mut assets,
-            &mut audio,
-            LoadingSession::from_request(LoadingRequest::generic_map_load(
-                "auto",
-                SkirmishSettings::default(),
-            )),
-            700,
-        );
-        let before_fallback_install = startup.clone();
-        assert_eq!(
-            retire_failed_loading_attempt(
-                &mut slot,
-                &mut startup,
-                &mut assets,
-                LoadingFailurePolicy::InstallGenericFallback
-            ),
-            LoadingFailurePolicy::InstallGenericFallback
-        );
-        assert_eq!(startup, before_fallback_install);
-        assert!(assets.is_available() && slot.is_none());
-        assert!(std::sync::Arc::ptr_eq(
-            &original.bytes,
-            &assets
-                .manager()
-                .unwrap()
-                .load_file_from_mix("sentinel.bin")
-                .unwrap()
-                .bytes
-        ));
-    }
-
-    /// Uses the production initial-map reader and real retail trig tables;
-    /// assertions distinguish initial read failure from context admission.
-    #[test]
-    #[ignore = "requires RA2_DIR with verified retail gamemd.exe"]
-    fn loading_preparation_consumes_real_source_and_returns_lease_on_initial_and_admission_failures()
-     {
-        let ra2_dir = PathBuf::from(std::env::var_os("RA2_DIR").expect("RA2_DIR"));
-        crate::map::retail_trig::install_from_dir(&ra2_dir);
-        assert!(crate::map::retail_trig::wave_tables_available());
-        let dir = std::env::temp_dir().join(format!(
-            "vera20k-loading-preparation-{}",
-            std::process::id()
-        ));
-        std::fs::create_dir_all(&dir).unwrap();
-        let map_path = dir.join("mp01t4.map");
-        let map_bytes = AssetManager::new(&ra2_dir)
-            .unwrap()
-            .get("Fight.MAP")
-            .expect("retail Fight.MAP fixture");
-        std::fs::write(&map_path, map_bytes).unwrap();
-        std::fs::write(dir.join("sentinel.bin"), b"keep this cache").unwrap();
-        let mut assets = crate::app::process_assets::ProcessAssets::from_startup(
-            Some(AssetManager::from_loose_root_for_test(&dir)),
-            None,
-            None,
-            None,
-        );
-        let original = assets
-            .manager()
-            .unwrap()
-            .load_file_from_mix("sentinel.bin")
-            .unwrap();
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.selected_map_file = Some(map_path.to_string_lossy().into_owned());
-        let request = LoadingRequest::unverified_legacy_skirmish(
-            launch,
-            unverified_seed(7),
-            SkirmishSettings::default(),
-        );
-        let mut session = LoadingSession::from_request(request);
-        session.job.asset_manager = assets.lease_for_loading();
-        let session =
-            prepare_loading_session(&mut assets, session, true, Some(ra2_dir.clone())).unwrap();
-        assert!(!session.first_frame_presented);
-        assert_eq!(
-            session.native.as_ref().unwrap().progress.current_value(),
-            0.0
-        );
-        let LoadingStage::Prepared(prepared) = &session.stage else {
-            panic!("prepared payload");
-        };
-        assert_eq!(
-            prepared.context.physical_source(),
-            prepared.initial.map_source()
-        );
-        assert_eq!(prepared.context.signed_new_ini_format(), 4);
-        assert!(!prepared.initial.map_data().waypoints.is_empty());
-        session.job.retire(&mut assets);
-
-        for (selected, expected_error) in [
-            (map_path.to_string_lossy().into_owned(), "Generic startup"),
-            (
-                dir.join("absent.map").to_string_lossy().into_owned(),
-                "absent.map",
-            ),
-        ] {
-            let mut session = LoadingSession::from_request(LoadingRequest::generic_map_load(
-                selected,
-                SkirmishSettings::default(),
-            ));
-            session.job.asset_manager = assets.lease_for_loading();
-            let err = prepare_loading_session(&mut assets, session, false, Some(ra2_dir.clone()))
-                .err()
-                .unwrap();
-            assert!(format!("{err:#}").contains(expected_error), "{err:#}");
-            assert!(assets.is_available() && !assets.is_leased());
-            assert!(std::sync::Arc::ptr_eq(
-                &original.bytes,
-                &assets
-                    .manager()
-                    .unwrap()
-                    .load_file_from_mix("sentinel.bin")
-                    .unwrap()
-                    .bytes
-            ));
-        }
-    }
-
-    #[test]
-    fn loading_side_comes_from_first_launch_node_country() {
-        let session = LoadingSession::from_request(LoadingRequest::unverified_legacy_skirmish(
-            test_launch_session(LaunchCountry::Korea),
-            unverified_seed(1),
-            SkirmishSettings::default(),
-        ));
-
-        assert_eq!(
-            session.native.as_ref().map(|native| native.variant),
-            Some(LoadingArtVariant::Alliance)
-        );
-    }
-
-    #[test]
-    fn loading_progress_row_snapshots_the_launch_player_name() {
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.player_name = "Commander".to_owned();
-        let session = LoadingSession::from_request(LoadingRequest::unverified_legacy_skirmish(
-            launch,
-            unverified_seed(22),
-            SkirmishSettings::default(),
-        ));
-
-        assert_eq!(
-            session
-                .native
-                .as_ref()
-                .map(|native| native.progress_row.label.as_str()),
-            Some("Commander"),
-        );
-    }
-
-    #[test]
-    fn loading_session_preserves_selected_map_filename() {
-        let session = LoadingSession::from_request(LoadingRequest::unverified_legacy_skirmish(
-            test_launch_session(LaunchCountry::Yuri),
-            unverified_seed(2),
-            SkirmishSettings::default(),
-        ));
-
-        assert_eq!(session.stage.request().selected_map_file(), "mp01t4.map");
-        assert_eq!(
-            session
-                .stage
-                .request()
-                .skirmish_launch_session()
-                .and_then(|launch| launch.selected_map_file.as_deref()),
-            Some("mp01t4.map")
-        );
-    }
-
-    #[test]
-    fn loading_session_selects_native_progress_cadence_from_map_kind() {
-        let selected = LoadingSession::from_request(LoadingRequest::unverified_legacy_skirmish(
-            test_launch_session(LaunchCountry::America),
-            unverified_seed(20),
-            SkirmishSettings::default(),
-        ));
-        assert_eq!(
-            selected
-                .native
-                .as_ref()
-                .map(|native| native.progress_cadence),
-            Some(NativeLoadingProgressCadence::SelectedMap)
-        );
-
-        let mut random_map = test_launch_session(LaunchCountry::America);
-        random_map.selected_map_file = Some("RandMap.Sed".to_string());
-        let random = LoadingSession::from_request(LoadingRequest::unverified_legacy_skirmish(
-            random_map,
-            unverified_seed(21),
-            SkirmishSettings::default(),
-        ));
-        assert_eq!(
-            random.native.as_ref().map(|native| native.progress_cadence),
-            Some(NativeLoadingProgressCadence::RandomMapHalved)
-        );
-    }
-
-    #[test]
-    fn gsi_04_12_loading_request_preview_is_presentation_only() {
-        use crate::map::waypoints::Waypoint;
-
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.selected_map_file = Some("RandMap.Sed".to_string());
-        let mut map =
-            crate::map::rmg::emit::empty_map_file(&crate::map::rmg::RmgOptions::default(), 32, 32);
-        map.waypoints.insert(
-            0,
-            Waypoint {
-                index: 0,
-                rx: 10,
-                ry: 20,
-            },
-        );
-        map.waypoints.insert(
-            1,
-            Waypoint {
-                index: 1,
-                rx: 30,
-                ry: 40,
-            },
-        );
-        let generated = crate::map::rmg::GeneratedMap {
-            map_file: map,
-            mapgen_continuation: crate::map::rmg::RmgRng::new(0x1234).into_continuation(),
-            construction_trace: crate::map::rmg::RmgConstructionTrace::default(),
-            start_waypoints: vec![(0, 10, 20), (1, 30, 40)],
-            stages_run: Vec::new(),
-            unfilled_start_slots: 0,
-        };
-        let request = LoadingRequest::unverified_legacy_skirmish(
-            launch.clone(),
-            unverified_seed(0x1234),
-            SkirmishSettings::default(),
-        )
-        .with_random_map_preview(Some(generated));
-
-        assert_eq!(
-            request
-                .random_map_preview()
-                .expect("loading request owns setup preview")
-                .start_waypoints,
-            vec![(0, 10, 20), (1, 30, 40)]
-        );
-        let assignments = selected_map_start_assignments(&launch, None);
-        assert!(
-            assignments.is_empty(),
-            "preview waypoints cannot resolve gameplay starts"
-        );
-    }
-
-    #[test]
-    fn gsi_04_12_generated_prefix_uses_accepted_staging_once() {
-        use crate::map::map_file::MapCell;
-
-        let selected = "RandMap.Sed";
-        let staged = [(0, 70, 70)];
-        let preview_waypoints = [(0, 110, 110), (1, 130, 110)];
-        let regenerated = [(0, 70, 90), (1, 90, 90)];
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.selected_map_file = Some(selected.to_string());
-        let accepted =
-            accepted_random_map_with_starts(selected, 0x1212, &preview_waypoints, &staged);
-        let mut regenerated_map = crate::map::rmg::emit::empty_map_file(
-            &crate::map::rmg::RmgOptions::default(),
-            100,
-            100,
-        );
-        regenerated_map
-            .waypoints
-            .extend(regenerated.iter().map(|&(slot, rx, ry)| {
-                (
-                    u32::from(slot),
-                    crate::map::waypoints::Waypoint {
-                        index: u32::from(slot),
-                        rx,
-                        ry,
-                    },
-                )
-            }));
-        regenerated_map.cells = [(60, 60), (60, 140), (140, 60), (140, 140)]
-            .into_iter()
-            .map(|(rx, ry)| MapCell {
-                rx,
-                ry,
-                tile_index: 0,
-                sub_tile: 0,
-                z: 0,
-            })
-            .collect();
-        let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-            regenerated_map,
-            crate::app::frontend::list_maps::LoadedMapSource::Generated {
-                seed_name: selected.to_ascii_lowercase(),
-            },
-        );
-        let request = LoadingRequest::unverified_legacy_skirmish(
-            launch.clone(),
-            unverified_seed(0x1212),
-            SkirmishSettings::default(),
-        )
-        .with_accepted_random_map(Some(accepted));
-
-        let prepared = request.prepare_initial(initial).unwrap();
-        let context = &prepared.context;
-        let initial = &prepared.initial;
-        let request = &prepared.request;
-        let projection = context.stock_offline_projection();
-        let final_starts = projection
-            .final_gathered_starts()
-            .iter()
-            .map(|waypoint| (waypoint.index as u8, waypoint.rx, waypoint.ry))
-            .collect::<Vec<_>>();
-        assert_eq!(final_starts.len(), 2);
-        assert_eq!(final_starts[0], staged[0]);
-        let gathered_fallback = final_starts[1];
-        assert!(
-            !staged.contains(&gathered_fallback)
-                && !preview_waypoints.contains(&gathered_fallback)
-                && !regenerated.contains(&gathered_fallback),
-            "deficient Gather must add a distinct temporary start"
-        );
-        assert_ne!(final_starts, preview_waypoints);
-        assert_ne!(final_starts, regenerated);
-        let active_starts = crate::map::waypoints::multiplayer_start_waypoints(
-            projection.active_scenario_waypoints(),
-        )
-        .into_iter()
-        .map(|waypoint| (waypoint.index as u8, waypoint.rx, waypoint.ry))
-        .collect::<Vec<_>>();
-        assert_eq!(active_starts, staged);
-
-        let assignments = selected_map_start_assignments(&launch, Some(projection));
-        let composition = build_random_map_loading_composition(
-            &launch,
-            None,
-            [800, 600],
-            Some(DecodedPreview {
-                width: 300,
-                height: 100,
-                rgba: vec![255; 300 * 100 * 4],
-            }),
-            initial.map_data(),
-            projection.active_scenario_waypoints(),
-            &assignments,
-        );
-        assert_eq!(
-            composition
-                .markers
-                .iter()
-                .map(|marker| {
-                    (
-                        marker.waypoint.index as u8,
-                        marker.waypoint.rx,
-                        marker.waypoint.ry,
-                    )
-                })
-                .collect::<Vec<_>>(),
-            staged,
-            "random loading markers read the active Scenario staging copy"
-        );
-
-        let staged_session = crate::app::loading::init::scenario_start_waypoints_for_load(
-            initial.map_data(),
-            Some(projection),
-        );
-        let regenerated_session =
-            crate::app::loading::init::scenario_start_waypoints_for_load(initial.map_data(), None);
-        assert_eq!(
-            staged_session
-                .iter()
-                .map(|(&index, &(rx, ry))| (index as u8, rx, ry))
-                .collect::<Vec<_>>(),
-            staged
-        );
-        assert_ne!(staged_session, regenerated_session);
-        let gathered_session = final_starts
-            .iter()
-            .map(|&(index, rx, ry)| (u32::from(index), (rx, ry)))
-            .collect();
-        assert_ne!(
-            staged_session, gathered_session,
-            "the active Scenario table excludes Gather-only fallback starts"
-        );
-        let staged_sim = crate::sim::world::Simulation::from_descriptor(
-            &crate::sim::scenario_session::ScenarioDescriptor {
-                seed: 0x1212,
-                map_width: 256,
-                map_height: 256,
-                mp_start_waypoints: staged_session,
-                ..Default::default()
-            },
-        );
-        let regenerated_sim = crate::sim::world::Simulation::from_descriptor(
-            &crate::sim::scenario_session::ScenarioDescriptor {
-                seed: 0x1212,
-                map_width: 256,
-                map_height: 256,
-                mp_start_waypoints: regenerated_session,
-                ..Default::default()
-            },
-        );
-        let gathered_sim = crate::sim::world::Simulation::from_descriptor(
-            &crate::sim::scenario_session::ScenarioDescriptor {
-                seed: 0x1212,
-                map_width: 256,
-                map_height: 256,
-                mp_start_waypoints: gathered_session,
-                ..Default::default()
-            },
-        );
-        assert_eq!(
-            staged_sim
-                .session
-                .mp_start_waypoints
-                .iter()
-                .map(|(&index, &(rx, ry))| (index as u8, rx, ry))
-                .collect::<Vec<_>>(),
-            staged
-        );
-        assert_ne!(staged_sim.state_hash(), gathered_sim.state_hash());
-        assert_ne!(staged_sim.state_hash(), regenerated_sim.state_hash());
-        assert!(
-            request.accepted_rmg_start_staging.is_none(),
-            "accepted setup staging transfers exactly once"
-        );
-        assert!(
-            request.random_map_preview().is_some(),
-            "presentation preview remains available to loading composition"
-        );
-    }
-
-    #[test]
-    fn load_descriptor_source_family_format_matrix() {
-        use crate::app::loading::fresh_scenario::{
-            FreshMapMaterialization, FreshScenarioFamily, FreshStartupProvenance,
-        };
-
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let signed_formats = [
-            (None, 0, false),
-            (Some(1), 1, false),
-            (Some(2), 2, true),
-            (Some(4), 4, true),
-            (Some(-7), -7, false),
-        ];
-        for source in [
-            crate::app::frontend::list_maps::LoadedMapSource::Loose {
-                path: std::path::PathBuf::from("mp01t4.map"),
-                payload_len: 17,
-            },
-            crate::app::frontend::list_maps::LoadedMapSource::Mix {
-                logical_name: "mp01t4.map".to_string(),
-                source_archive: "mapsmd03.mix".to_string(),
-                entry_id: 0x1234,
-                payload_len: 19,
-            },
-        ] {
-            for (format, expected_signed, expected_pack_gate) in signed_formats {
-                let mut map = prefix_test_map(&starts);
-                map.basic.new_ini_format = format;
-                let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-                    map,
-                    source.clone(),
-                );
-                let request = LoadingRequest::unverified_legacy_skirmish(
-                    test_launch_session(LaunchCountry::America),
-                    unverified_seed(0x1A2B_3C4D),
-                    SkirmishSettings::default(),
-                );
-                let prepared = request
-                    .prepare_initial(initial)
-                    .expect("authored Loose/MIX stock context");
-                let context = &prepared.context;
-                let request = &prepared.request;
-                assert_eq!(context.physical_source(), &source);
-                assert_eq!(context.materialization(), FreshMapMaterialization::Authored);
-                assert_eq!(context.family(), FreshScenarioFamily::StockOffline);
-                assert_eq!(
-                    context.startup_provenance(),
-                    FreshStartupProvenance::ResolvedLegacy
-                );
-                assert_eq!(context.match_seed(), 0x1A2B_3C4D);
-                assert_eq!(context.signed_new_ini_format(), expected_signed);
-                context
-                    .validate_terminal_transfer(request.startup(), &source, expected_signed)
-                    .expect("the prepared bundle retains matching terminal inputs");
-                assert_eq!(
-                    context.authored_pack_bodies_enabled(),
-                    expected_pack_gate,
-                    "only signed NewINIFormat > 1 gates authored pack bodies"
-                );
-            }
-        }
-
-        for mode_id in [1, 2] {
-            let selected = format!("Accepted{mode_id}.SED");
-            let mut launch = test_launch_session(LaunchCountry::America);
-            launch.selected_map_file = Some(selected.clone());
-            if mode_id == 2 {
-                launch.mode = SkirmishLaunchMode {
-                    id: 2,
-                    ui_name_key: "GUI:FreeForAll".to_string(),
-                    tooltip_key: "STT:ModeFreeForAll".to_string(),
-                    override_file: "MPFreeForAllMD.ini".to_string(),
-                    map_filter: "standard".to_string(),
-                    random_maps_allowed: true,
-                    allies_allowed: false,
-                    must_ally: false,
-                };
-            }
-            let accepted = accepted_random_map_with_starts(&selected, 0x2345, &starts, &starts);
-            let mut map = prefix_test_map(&starts);
-            map.basic.new_ini_format = Some(4);
-            let source = crate::app::frontend::list_maps::LoadedMapSource::Generated {
-                seed_name: selected.to_ascii_lowercase(),
-            };
-            let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-                map,
-                source.clone(),
-            );
-            let request = LoadingRequest::unverified_legacy_skirmish(
-                launch,
-                unverified_seed(0x2345),
-                SkirmishSettings::default(),
-            )
-            .with_accepted_random_map(Some(accepted));
-            let prepared = request
-                .prepare_initial(initial)
-                .expect("accepted Battle/FFA generated context");
-            let context = &prepared.context;
-            let request = &prepared.request;
-            assert_eq!(context.physical_source(), &source);
-            assert_eq!(
-                context.materialization(),
-                FreshMapMaterialization::AcceptedGenerated
-            );
-            assert_eq!(context.signed_new_ini_format(), 4);
-            context
-                .validate_terminal_transfer(request.startup(), &source, 4)
-                .expect("generated terminal owners still agree");
-            assert!(
-                !context.authored_pack_bodies_enabled(),
-                "serialized format cannot turn generated materialization into authored Mark"
-            );
-        }
-    }
-
-    #[test]
-    fn accepted_and_resolved_legacy_share_one_stock_cursor_shape() {
-        use crate::app::loading::fresh_scenario::{FreshScenarioFamily, FreshStartupProvenance};
-
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let source = crate::app::frontend::list_maps::LoadedMapSource::Loose {
-            path: std::path::PathBuf::from("mp01t4.map"),
-            payload_len: 23,
-        };
-        let mut map = prefix_test_map(&starts);
-        map.basic.new_ini_format = Some(4);
-        let initial =
-            crate::app::loading::init::MapLoadInitial::from_test_map_source(map, source.clone());
-        let mut legacy_map = prefix_test_map(&starts);
-        legacy_map.basic.new_ini_format = Some(4);
-        let legacy_initial =
-            crate::app::loading::init::MapLoadInitial::from_test_map_source(legacy_map, source);
-        let seed = 0x3456_789A;
-        let mut next = 1;
-        let prepared = prepared_startup(&mut next, seed);
-        let legacy_session = prepared.session.launch_session().clone();
-        let accepted = LoadingRequest::accepted_skirmish(prepared, SkirmishSettings::default());
-        let resolved_legacy = LoadingRequest::unverified_legacy_skirmish(
-            legacy_session,
-            unverified_seed(seed),
-            SkirmishSettings::default(),
-        );
-        let accepted = accepted.prepare_initial(initial).unwrap();
-        let resolved_legacy = resolved_legacy.prepare_initial(legacy_initial).unwrap();
-        let accepted_context = &accepted.context;
-        let legacy_context = &resolved_legacy.context;
-        assert_eq!(accepted_context.family(), FreshScenarioFamily::StockOffline);
-        assert_eq!(accepted_context.family(), legacy_context.family());
-        assert_eq!(accepted_context.match_seed(), legacy_context.match_seed());
-        assert_eq!(
-            accepted_context
-                .stock_offline_projection()
-                .final_gathered_starts(),
-            legacy_context
-                .stock_offline_projection()
-                .final_gathered_starts()
-        );
-        assert_eq!(
-            accepted_context.stock_offline_projection().start_table(),
-            legacy_context.stock_offline_projection().start_table()
-        );
-        assert_eq!(
-            accepted_context.startup_provenance(),
-            FreshStartupProvenance::Accepted
-        );
-        assert_eq!(
-            legacy_context.startup_provenance(),
-            FreshStartupProvenance::ResolvedLegacy
-        );
-
-        let accepted_parts = accepted.context.into_stock_offline_parts();
-        let legacy_parts = resolved_legacy.context.into_stock_offline_parts();
-        let mut accepted_owner = crate::sim::scenario_bootstrap::ScenarioBootstrapRng::new(seed);
-        let mut legacy_owner = crate::sim::scenario_bootstrap::ScenarioBootstrapRng::new(seed);
-        let _ = accepted_owner
-            .install_pre_fill_scenario_prefix_plan(accepted_parts.scenario_prefix)
-            .unwrap();
-        let _ = legacy_owner
-            .install_pre_fill_scenario_prefix_plan(legacy_parts.scenario_prefix)
-            .unwrap();
-        assert_eq!(
-            accepted_owner.logical_states_for_test(),
-            legacy_owner.logical_states_for_test(),
-            "startup provenance cannot change the verified stock prefix cursor"
-        );
-    }
-
-    #[test]
-    fn generic_manual_and_unresolved_legacy_reject_before_receipt_or_staging() {
-        let selected = "Rejected.SED";
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let source = crate::app::frontend::list_maps::LoadedMapSource::Generated {
-            seed_name: selected.to_string(),
-        };
-        let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-            prefix_test_map(&starts),
-            source,
-        );
-        let accepted = accepted_random_map_with_starts(selected, 0x4567, &starts, &starts);
-        let mut generic = LoadingRequest::generic_map_load(selected, SkirmishSettings::default())
-            .with_accepted_random_map(Some(accepted));
-        let generic_err = generic.admit_context(&initial).unwrap_err();
-        assert!(format!("{generic_err:#}").contains("Generic startup"));
-        assert!(generic.accepted_rmg_start_staging.is_some());
-
-        let accepted = accepted_random_map_with_starts(selected, 0x4567, &starts, &starts);
-        let mut unresolved_session = test_launch_session(LaunchCountry::America);
-        unresolved_session.selected_map_file = Some(selected.to_string());
-        unresolved_session.local.country_random = true;
-        let mut unresolved = LoadingRequest::unverified_legacy_skirmish(
-            unresolved_session,
-            unverified_seed(0x4567),
-            SkirmishSettings::default(),
-        )
-        .with_accepted_random_map(Some(accepted));
-        let unresolved_err = unresolved.admit_context(&initial).unwrap_err();
-        assert!(format!("{unresolved_err:#}").contains("local slot still has a random country"));
-        assert!(unresolved.accepted_rmg_start_staging.is_some());
-
-        let authored = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-            prefix_test_map(&starts),
-            crate::app::frontend::list_maps::LoadedMapSource::Loose {
-                path: std::path::PathBuf::from("manual.map"),
-                payload_len: 1,
-            },
-        );
-        let mut manual_session = test_launch_session(LaunchCountry::America);
-        manual_session.selected_map_file = Some(" auto ".to_string());
-        let mut manual = LoadingRequest::unverified_legacy_skirmish(
-            manual_session,
-            unverified_seed(0x4567),
-            SkirmishSettings::default(),
-        );
-        let manual_err = manual.admit_context(&authored).unwrap_err();
-        assert!(format!("{manual_err:#}").contains("no exact selected map record"));
-    }
-
-    #[test]
-    fn loading_assignments_keep_gather_table_slots_separate_from_sparse_waypoint_indices() {
-        use crate::app::loading::composition::{
-            PreviewAspectFit, ProjectedPlayfieldBounds, build_mmpb_marker_records,
-            native_loading_waypoint_prefix,
-        };
-
-        let mut launch = test_launch_session(LaunchCountry::America);
-        let mut second_opponent = launch.opponents[0].clone();
-        second_opponent.color_index = 2;
-        second_opponent.start_position = LaunchStartPosition::Position(2);
-        launch.opponents.push(second_opponent);
-        launch.pre_fill_house_roster =
-            crate::skirmish_launch::PreFillHouseRoster::from_compact_skirmish(2);
-
-        // Gather target 3 retains raw slots 0 and 2, then appends a fallback
-        // at vector position 2. The retained slot and fallback consequently
-        // carry the same Waypoint::index even though the assignment table has
-        // three distinct positions.
-        let map = prefix_test_map(&[(0, 20, 24), (2, 42, 46), (3, 54, 58)]);
-        let descriptor =
-            crate::sim::scenario_bootstrap::MatchLaunchDescriptor::from_resolved(launch.clone())
-                .unwrap();
-        let plan = crate::sim::scenario_bootstrap::prepare_stock_offline_scenario_prefix_plan(
-            &descriptor,
-            &map,
-            &map.waypoints,
-            0x1223,
-        )
-        .unwrap();
-        assert_eq!(
-            plan.final_gathered_starts()
-                .iter()
-                .map(|waypoint| waypoint.index)
-                .collect::<Vec<_>>(),
-            vec![0, 2, 2]
-        );
-
-        let assignments = selected_map_start_assignments(&launch, Some(plan.projection()));
-        assert_eq!(
-            assignments
-                .iter()
-                .map(|assignment| (assignment.start_index, assignment.participant))
-                .collect::<Vec<_>>(),
-            vec![
-                (0, LoadingParticipantId::Local),
-                (1, LoadingParticipantId::Opponent(0)),
-                (2, LoadingParticipantId::Opponent(1)),
-            ],
-            "loading assignments are keyed by Scenario start-table position"
-        );
-
-        // The compositor still reads original Scenario geometry. Native's odd
-        // sparse-prefix rule visits raw waypoint 2 and colors it from table[2],
-        // not from the retained Gather vector entry at position 1.
-        let raw_prefix = native_loading_waypoint_prefix(&map.waypoints);
-        let markers = build_mmpb_marker_records(
-            &raw_prefix,
-            &assignments,
-            ProjectedPlayfieldBounds {
-                min_x: 0,
-                min_y: 0,
-                extent_x: 1_000,
-                extent_y: 1_000,
-            },
-            MmpbRegionRect {
-                x: 0,
-                y: 0,
-                width: 200,
-                height: 200,
-            },
-            PreviewAspectFit {
-                scale_1000: 1_000,
-                width: 200,
-                height: 200,
-                pad_x: 0,
-                pad_y: 0,
-            },
-        );
-        assert_eq!(markers.len(), 2);
-        assert_eq!(markers[1].waypoint.index, 2);
-        assert_eq!(markers[1].participant, LoadingParticipantId::Opponent(1));
-    }
-
-    #[test]
-    fn gsi_04_12_generated_prefix_rejects_presentation_only_preview() {
-        let selected = "RandMap.Sed";
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.selected_map_file = Some(selected.to_string());
-        let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-            prefix_test_map(&starts),
-            crate::app::frontend::list_maps::LoadedMapSource::Generated {
-                seed_name: selected.to_string(),
-            },
-        );
-        let mut request = LoadingRequest::unverified_legacy_skirmish(
-            launch,
-            unverified_seed(0x1313),
-            SkirmishSettings::default(),
-        )
-        .with_random_map_preview(Some(generated_preview_with_starts(0x1313, &starts)));
-
-        let err = request.admit_context(&initial).unwrap_err();
-        assert!(
-            format!("{err:#}").contains("no accepted setup start staging"),
-            "unexpected error: {err:#}"
-        );
-    }
-
-    #[test]
-    fn generated_prefix_rejects_mismatched_source_name() {
-        let selected = "RandMap.Sed";
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.selected_map_file = Some(selected.to_string());
-        let accepted = accepted_random_map_with_starts(selected, 0x1414, &starts, &starts);
-        let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-            prefix_test_map(&starts),
-            crate::app::frontend::list_maps::LoadedMapSource::Generated {
-                seed_name: "Other.Sed".to_string(),
-            },
-        );
-        let mut request = LoadingRequest::unverified_legacy_skirmish(
-            launch,
-            unverified_seed(0x1414),
-            SkirmishSettings::default(),
-        )
-        .with_accepted_random_map(Some(accepted));
-
-        let err = request.admit_context(&initial).unwrap_err();
-        assert!(
-            format!("{err:#}").contains("does not match selected record"),
-            "unexpected error: {err:#}"
-        );
-    }
-
-    #[test]
-    fn generated_prefix_rejects_cooperative_mode() {
-        let selected = "RandMap.Sed";
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.selected_map_file = Some(selected.to_string());
-        let cooperative = crate::skirmish_modes::stock_skirmish_modes()
-            .into_iter()
-            .find(|mode| mode.id == 3)
-            .expect("retail Cooperative row");
-        launch.mode = SkirmishLaunchMode::from_game_mode(&cooperative);
-        let accepted = accepted_random_map_with_starts(selected, 0x1515, &starts, &starts);
-        let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-            prefix_test_map(&starts),
-            crate::app::frontend::list_maps::LoadedMapSource::Generated {
-                seed_name: selected.to_string(),
-            },
-        );
-        let mut request = LoadingRequest::unverified_legacy_skirmish(
-            launch,
-            unverified_seed(0x1515),
-            SkirmishSettings::default(),
-        )
-        .with_accepted_random_map(Some(accepted));
-
-        let err = request.admit_context(&initial).unwrap_err();
-        assert!(
-            format!("{err:#}").contains("unsupported for stock mode id 3"),
-            "unexpected error: {err:#}"
-        );
-        assert!(
-            request.accepted_rmg_start_staging.is_some(),
-            "unsupported family rejects before consuming accepted staging"
-        );
-    }
-
-    #[test]
-    fn generated_prefix_rejects_a_spoofed_stock_row_before_consuming_staging() {
-        let selected = "RandMap.SED";
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.selected_map_file = Some(selected.to_string());
-        launch.mode.tooltip_key = "STT:SpoofedBattle".to_string();
-        let accepted = accepted_random_map_with_starts(selected, 0x1516, &starts, &starts);
-        let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-            prefix_test_map(&starts),
-            crate::app::frontend::list_maps::LoadedMapSource::Generated {
-                seed_name: selected.to_string(),
-            },
-        );
-        let mut request = LoadingRequest::unverified_legacy_skirmish(
-            launch,
-            unverified_seed(0x1516),
-            SkirmishSettings::default(),
-        )
-        .with_accepted_random_map(Some(accepted));
-
-        let err = request.admit_context(&initial).unwrap_err();
-        assert!(format!("{err:#}").contains("not the validated active-retail stock row"));
-        assert!(request.accepted_rmg_start_staging.is_some());
-    }
-
-    #[test]
-    fn authored_prefix_rejects_random_map_staging_for_loose_and_mix_sources() {
-        let selected = "mp01t4.map";
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let sources = [
-            crate::app::frontend::list_maps::LoadedMapSource::Loose {
-                path: std::path::PathBuf::from(selected),
-                payload_len: 1,
-            },
-            crate::app::frontend::list_maps::LoadedMapSource::Mix {
-                logical_name: selected.to_string(),
-                source_archive: "mapsmd03.mix".to_string(),
-                entry_id: 7,
-                payload_len: 1,
-            },
-        ];
-        for source in sources {
-            let mut launch = test_launch_session(LaunchCountry::America);
-            launch.selected_map_file = Some(selected.to_string());
-            let accepted = accepted_random_map_with_starts(selected, 0x1616, &starts, &starts);
-            let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-                prefix_test_map(&starts),
-                source,
-            );
-            let mut request = LoadingRequest::unverified_legacy_skirmish(
-                launch,
-                unverified_seed(0x1616),
-                SkirmishSettings::default(),
-            )
-            .with_accepted_random_map(Some(accepted));
-
-            let err = request.admit_context(&initial).unwrap_err();
-            assert!(
-                format!("{err:#}").contains("cannot attach to an authored map source"),
-                "unexpected error: {err:#}"
-            );
-        }
-    }
-
-    #[test]
-    fn stock_prefix_rejects_legacy_fallback_source() {
-        let starts = [(0, 20, 24), (1, 42, 46)];
-        let launch = test_launch_session(LaunchCountry::America);
-        let initial = crate::app::loading::init::MapLoadInitial::from_test_map_source(
-            prefix_test_map(&starts),
-            crate::app::frontend::list_maps::LoadedMapSource::LegacyFallback {
-                label: "fixture".to_string(),
-            },
-        );
-        let mut request = LoadingRequest::unverified_legacy_skirmish(
-            launch,
-            unverified_seed(0x1717),
-            SkirmishSettings::default(),
-        );
-
-        let err = request.admit_context(&initial).unwrap_err();
-        assert!(
-            format!("{err:#}")
-                .contains("requires an exact Loose, MIX, or accepted generated source"),
-            "unexpected error: {err:#}"
-        );
-
-        let mut next = 1;
-        let mut accepted = LoadingRequest::accepted_skirmish(
-            prepared_startup(&mut next, 0x1717),
-            SkirmishSettings::default(),
-        );
-        let accepted_err = accepted.admit_context(&initial).unwrap_err();
-        assert!(
-            format!("{accepted_err:#}")
-                .contains("requires an exact Loose, MIX, or accepted generated source"),
-            "unexpected error: {accepted_err:#}"
-        );
-    }
-
-    #[test]
-    fn gsi_04_12_stock_ffa_preview_can_only_supply_loading_fallback_pixels() {
-        use crate::map::map_file::MapCell;
-        use crate::map::waypoints::Waypoint;
-
-        let mut launch = test_launch_session(LaunchCountry::America);
-        launch.mode = SkirmishLaunchMode {
-            id: 2,
-            ui_name_key: "GUI:FreeForAll".to_string(),
-            tooltip_key: "STT:ModeFreeForAll".to_string(),
-            override_file: "MPFreeForAllMD.ini".to_string(),
-            map_filter: "standard".to_string(),
-            random_maps_allowed: true,
-            allies_allowed: false,
-            must_ally: false,
-        };
-        launch.selected_map_file = Some("RandMap.Sed".to_string());
-
-        let mut map = crate::map::rmg::emit::empty_map_file(
-            &crate::map::rmg::RmgOptions::default(),
-            100,
-            100,
-        );
-        map.cells = vec![
-            MapCell {
-                rx: 60,
-                ry: 60,
-                tile_index: 0,
-                sub_tile: 0,
-                z: 0,
-            },
-            MapCell {
-                rx: 60,
-                ry: 140,
-                tile_index: 0,
-                sub_tile: 0,
-                z: 0,
-            },
-            MapCell {
-                rx: 140,
-                ry: 60,
-                tile_index: 0,
-                sub_tile: 0,
-                z: 0,
-            },
-        ];
-        map.waypoints.insert(
-            0,
-            Waypoint {
-                index: 0,
-                rx: 70,
-                ry: 70,
-            },
-        );
-        map.waypoints.insert(
-            1,
-            Waypoint {
-                index: 1,
-                rx: 90,
-                ry: 70,
-            },
-        );
-        map.waypoints.insert(
-            2,
-            Waypoint {
-                index: 2,
-                rx: 70,
-                ry: 90,
-            },
-        );
-        let mut launch_map = crate::map::rmg::emit::empty_map_file(
-            &crate::map::rmg::RmgOptions::default(),
-            100,
-            100,
-        );
-        launch_map.cells = map.cells.clone();
-        let active_scenario_waypoints = map.waypoints.clone();
-        let generated = crate::map::rmg::GeneratedMap {
-            map_file: map,
-            mapgen_continuation: crate::map::rmg::RmgRng::new(0x4567).into_continuation(),
-            construction_trace: crate::map::rmg::RmgConstructionTrace::default(),
-            start_waypoints: vec![(0, 70, 70), (1, 90, 70), (2, 70, 90)],
-            stages_run: Vec::new(),
-            unfilled_start_slots: 0,
-        };
-        let _request = LoadingRequest::unverified_legacy_skirmish(
-            launch.clone(),
-            unverified_seed(0x4567),
-            SkirmishSettings::default(),
-        )
-        .with_random_map_preview(Some(generated));
-        let assignments = selected_map_start_assignments(&launch, None);
-        assert!(
-            assignments.is_empty(),
-            "only the launch-time `.SED` regeneration may resolve participants"
-        );
-        let preview = DecodedPreview {
-            width: 300,
-            height: 100,
-            rgba: vec![255; 300 * 100 * 4],
-        };
-        let composition = build_random_map_loading_composition(
-            &launch,
-            None,
-            [800, 600],
-            Some(preview),
-            &launch_map,
-            &active_scenario_waypoints,
-            &assignments,
-        );
-        let prepared = composition.preview.expect("FFA retained preview");
-        assert_eq!(
-            prepared
-                .image
-                .rgba
-                .chunks_exact(4)
-                .filter(|pixel| *pixel == [0, 0, 0, 255])
-                .count(),
-            3 * 4 * 4,
-            "all three valid FFA starts are burned black"
-        );
-        assert!(composition.markers.is_empty());
-    }
-
-    #[test]
-    fn random_map_progress_uses_native_integer_halving_and_raw_200_terminal() {
-        let cadence = NativeLoadingProgressCadence::RandomMapHalved;
-
-        assert_eq!(cadence.effective_percent(3), 1);
-        assert_eq!(cadence.effective_percent(90), 45);
-        assert_eq!(
-            cadence.effective_percent(cadence.terminal_raw_percent()),
-            100
-        );
-        assert_eq!(
-            NativeLoadingProgressCadence::SelectedMap.effective_percent(3),
-            3
-        );
-        assert_eq!(
-            NativeLoadingProgressCadence::SelectedMap.terminal_raw_percent(),
-            100
-        );
-    }
-
-    #[test]
-    fn loading_session_falls_back_without_native_session_only_outside_parity_path() {
-        let session = LoadingSession::from_request(LoadingRequest::generic_map_load(
-            "auto",
-            SkirmishSettings::default(),
-        ));
-
-        assert!(session.native.is_none());
-        assert!(session.stage.request().skirmish_launch_session().is_none());
-        assert_eq!(session.stage.request().selected_map_file(), "auto");
-    }
-
-    #[test]
-    fn loading_session_starts_at_initial_map_selection_phase() {
-        let session = LoadingSession::from_request(LoadingRequest::unverified_legacy_skirmish(
-            test_launch_session(LaunchCountry::America),
-            unverified_seed(3),
-            SkirmishSettings::default(),
-        ));
-
-        assert!(matches!(session.stage, LoadingStage::Selected(_)));
-    }
-
-    #[test]
-    fn loading_request_moves_exact_startup_authority_once() {
-        struct Clock;
-        impl crate::match_bootstrap::MatchSeedClock for Clock {
-            fn low_u32(&mut self) -> u32 {
-                0x1234_5678
-            }
-
-            fn source(&self) -> crate::match_bootstrap::MatchSeedSource {
-                crate::match_bootstrap::MatchSeedSource::Controlled
-            }
-
-            fn seed_authority_certifying(&self) -> bool {
-                true
-            }
-        }
-
-        let session = test_launch_session(LaunchCountry::America);
-        let accepted = match crate::match_bootstrap::classify_startup_session(&session) {
-            crate::match_bootstrap::StartupSessionClassification::AcceptedExplicitFixedBattle(
-                accepted,
-            ) => accepted,
-            other => panic!("fixture was not accepted: {other:?}"),
-        };
-        let mut next = 1;
-        let correlation = crate::match_bootstrap::allocate_match_correlation(&mut next).unwrap();
-        let mut clock = Clock;
-        let prepared =
-            crate::match_bootstrap::prepare_match_startup(correlation, accepted, &mut clock);
-        let request =
-            LoadingRequest::accepted_skirmish(prepared.clone(), SkirmishSettings::default());
-
-        assert_eq!(request.startup().accepted(), Some(&prepared));
-        let initial = MapLoadInitial::from_test_map_source(
-            prefix_test_map(&[(0, 20, 24), (1, 42, 46)]),
-            crate::app::frontend::list_maps::LoadedMapSource::Loose {
-                path: PathBuf::from("mp01t4.map"),
-                payload_len: 1,
-            },
-        );
-        let loaded = request.prepare_initial(initial).unwrap();
-        assert_eq!(loaded.request.startup, LoadingStartup::Accepted(prepared));
-        assert_eq!(
-            loaded.context.physical_source(),
-            loaded.initial.map_source()
-        );
-    }
-
-    #[test]
-    fn replacing_loading_startup_retires_prior_admission() {
-        use crate::app::match_runtime::startup::MatchStartup;
-        let mut next = 1;
-        let prior = prepared_startup(&mut next, 0x1111_2222);
-        let replacement = prepared_startup(&mut next, 0x3333_4444);
-        let mut authority = MatchStartup::default();
-        assert!(authority.admits_ordinary_tick());
-        assert!(!authority.admits_exact_step());
-        authority.begin(Some(prior.correlation));
-        let simulation = crate::sim::world::Simulation::with_seed(u64::from(prior.seed.value));
-        authority
-            .acknowledge(prior.clone(), Some(&simulation), true, false)
-            .unwrap();
-        assert_eq!(authority.receipt(), Some(&receipt_for(&prior)));
-        assert!(authority.admits_exact_step());
-
-        // New loading attempt cannot expose the previous pair to captures.
-        authority.begin(Some(replacement.correlation));
-        assert!(authority.accepted().is_none());
-        assert!(!authority.admits_exact_step());
-        let baseline = authority.clone();
-        assert!(
-            authority
-                .acknowledge(prior, Some(&simulation), true, false)
-                .is_err()
-        );
-        assert_eq!(authority, baseline, "stale completion cannot install");
-        let simulation =
-            crate::sim::world::Simulation::with_seed(u64::from(replacement.seed.value));
-        authority
-            .acknowledge(replacement.clone(), Some(&simulation), true, false)
-            .unwrap();
-        assert_eq!(authority.startup(), Some(&replacement));
-        let accepted = authority.clone();
-        assert!(
-            authority
-                .acknowledge(replacement, Some(&simulation), true, false)
-                .is_err()
-        );
-        assert_eq!(
-            authority, accepted,
-            "duplicate completion cannot overwrite evidence"
-        );
-        authority.clear();
-        assert!(authority.accepted().is_none());
-        assert!(authority.admits_ordinary_tick());
-        assert!(!authority.admits_exact_step());
-    }
-
-    #[test]
-    fn rejected_startup_observation_does_not_publish_partial_admission() {
-        use crate::app::match_runtime::startup::MatchStartup;
-        let mut next = 1;
-        let prepared = prepared_startup(&mut next, 0x5555_6666);
-        let mut authority = MatchStartup::default();
-        authority.begin(Some(prepared.correlation));
-        let baseline = authority.clone();
-        let simulation = crate::sim::world::Simulation::with_seed(u64::from(prepared.seed.value));
-        for (sim, loading, spawn) in [
-            (None, true, false),
-            (Some(&simulation), false, false),
-            (Some(&simulation), true, true),
-        ] {
-            assert!(
-                authority
-                    .acknowledge(prepared.clone(), sim, loading, spawn)
-                    .is_err()
-            );
-            assert_eq!(authority, baseline);
-            assert!(!authority.admits_exact_step());
-        }
-        authority.clear();
-        assert!(authority.accepted().is_none());
-        authority.begin(None);
-        assert!(authority.admits_ordinary_tick());
-        assert!(!authority.admits_exact_step());
-    }
-
-    #[test]
-    fn loading_progress_standard_skirmish_initializes_one_lane_max_100() {
-        let progress = LoadingProgressState::standard_skirmish();
-
-        assert_eq!(progress.max_value(), 100.0);
-        assert_eq!(progress.current_value(), 0.0);
-        assert_eq!(progress.current_percent(), 0.0);
-    }
-
-    #[test]
-    fn loading_progress_duplicate_milestones_do_not_redraw() {
-        let mut progress = LoadingProgressState::standard_skirmish();
-
-        assert!(progress.advance_progress(3));
-        assert!(!progress.advance_progress(3));
-        assert_eq!(progress.current_value(), 3.0);
-    }
-
-    #[test]
-    fn loading_progress_lower_milestone_does_not_redraw() {
-        let mut progress = LoadingProgressState::standard_skirmish();
-
-        assert!(progress.advance_progress(8));
-        assert!(!progress.advance_progress(6));
-        assert_eq!(progress.current_value(), 8.0);
-    }
-
-    #[test]
-    fn loading_progress_advancing_milestone_requests_redraw() {
-        let mut progress = LoadingProgressState::standard_skirmish();
-
-        assert!(progress.advance_progress(3));
-        assert!(progress.advance_progress(8));
-        assert_eq!(progress.current_value(), 8.0);
-    }
-
-    #[test]
-    fn loading_progress_clipped_width_matches_native_formula_for_exact_values() {
-        let mut progress = LoadingProgressState::standard_skirmish();
-
-        assert_eq!(progress.fill_width_gamemd_ftol_positive_domain(326), 0);
-        assert!(progress.advance_progress(50));
-        assert_eq!(progress.fill_width_gamemd_ftol_positive_domain(326), 163);
-        assert!(progress.advance_progress(100));
-        assert_eq!(progress.fill_width_gamemd_ftol_positive_domain(326), 326);
-    }
-
-    #[test]
-    fn loading_progress_fill_width_uses_gamemd_ftol_positive_domain() {
-        let mut progress = LoadingProgressState::standard_skirmish();
-
-        assert!(progress.advance_progress(25));
-        assert_eq!(progress.fill_width_gamemd_ftol_positive_domain(400), 100);
-    }
-
-    #[test]
-    fn loading_progress_suppresses_nonadvancing_raw_native_calls() {
-        let mut progress = LoadingProgressState::standard_skirmish();
-
-        assert!(progress.advance_progress(8));
-        assert!(!progress.advance_progress(6));
-        assert!(progress.advance_progress(60));
-        assert!(!progress.advance_progress(58));
-        assert!(!progress.advance_progress(60));
-    }
-
-    #[test]
-    fn recording_sink_emits_full_monotonic_ledger_in_our_execution_order() {
-        // The values the loaders emit, in the order our pipeline crosses them
-        // (3 at present, 100 at Finished are emitted by the pump, not the loaders).
-        let mut loader_emits = vec![8, 12];
-        loader_emits.extend(theater_ramp_changed_values(42));
-        loader_emits.extend([
-            30, 31, 35, 45, 50, 55, 58, 60, 63, 65, 67, 68, 69, 70, 72, 74, 76, 78, 82, 86, 90, 93,
-            96, 98,
-        ]);
-        let mut sink = RecordingProgressSink::standard();
-        sink.milestone(3); // present handoff
-        for v in loader_emits.iter().copied() {
-            sink.milestone(v);
-        }
-        sink.milestone(100); // pump Finished
-
-        // Every emitted value advanced the bar (no suppressed/duplicate values),
-        // starts at 3, ends at 100, and is strictly monotonic.
-        assert_eq!(sink.emitted.first(), Some(&3));
-        assert_eq!(sink.emitted.last(), Some(&100));
-        assert_eq!(sink.emitted.len(), loader_emits.len() + 2);
-        assert!(sink.emitted.windows(2).all(|w| w[0] < w[1]));
-    }
-
-    #[test]
-    fn recording_sink_suppresses_nonadvancing_and_duplicate_milestones() {
-        let mut sink = RecordingProgressSink::standard();
-        // A stray lower/raw value (e.g. gamemd's raw 6 after 8) and duplicates
-        // must not advance the bar.
-        for v in [8, 6, 8, 12, 12, 30] {
-            sink.milestone(v);
-        }
-        assert_eq!(sink.emitted, vec![8, 12, 30]);
-    }
-
-    #[test]
-    fn loading_progress_standard_skirmish_selected_map_emits_verified_milestone_ledger() {
-        let ramp = theater_ramp_changed_values(42);
-        let mut expected = Vec::from([3, 8, 12]);
-        expected.extend(ramp);
-        expected.extend([
-            30, 31, 35, 45, 50, 55, 58, 60, 63, 65, 67, 68, 69, 70, 72, 74, 76, 78, 82, 86, 90, 93,
-            96, 98, 100,
-        ]);
-
-        let mut progress = LoadingProgressState::standard_skirmish();
-        let emitted: Vec<u32> = expected
-            .iter()
-            .copied()
-            .filter(|value| progress.advance_progress(*value))
-            .collect();
-
-        assert_eq!(emitted, expected);
-    }
-
-    #[test]
-    fn loading_progress_theater_ramp_stock_rulesmd_count_emits_13_through_25() {
-        let emitted = theater_ramp_changed_values(42);
-
-        assert_eq!(emitted, (13..=25).collect::<Vec<_>>());
-    }
-
-    #[test]
-    fn loading_progress_theater_ramp_nonmultiple_base_count_uses_native_quotient() {
-        let emitted = theater_ramp_changed_values(38);
-
-        assert_eq!(emitted, (13..=25).collect::<Vec<_>>());
-    }
-
-    #[test]
-    fn loading_progress_theater_ramp_zero_or_invalid_small_count_has_no_dynamic_values() {
-        assert!(theater_ramp_changed_values(0).is_empty());
-        assert!(theater_ramp_changed_values(12).is_empty());
-    }
-
-    #[test]
-    fn loading_theater_cache_mismatch_covers_first_same_and_changed_cases() {
-        assert!(theater_cache_mismatch(false, "TEMPERATE", "TEMPERATE"));
-        assert!(!theater_cache_mismatch(true, "TEMPERATE", "temperate"));
-        assert!(theater_cache_mismatch(true, "TEMPERATE", "SNOW"));
-    }
-
-    #[test]
-    fn loading_progress_read_ini_basic_milestones_precede_map_pack_milestones() {
-        let sequence = STANDARD_SKIRMISH_SELECTED_MAP_MILESTONES_AFTER_FIRST_RENDER;
-        let pos_55 = sequence.iter().position(|value| *value == 55).unwrap();
-        let pos_60 = sequence.iter().position(|value| *value == 60).unwrap();
-        let pos_63 = sequence.iter().position(|value| *value == 63).unwrap();
-
-        assert!(pos_55 < pos_63);
-        assert!(pos_60 < pos_63);
-    }
-
-    #[test]
-    fn selected_native_gate_preserves_metadata_and_raw_progress_cadence() {
-        for (cadence, expected) in [
-            (NativeLoadingProgressCadence::SelectedMap, 12.0),
-            (NativeLoadingProgressCadence::RandomMapHalved, 6.0),
-        ] {
-            let mut native = NativeLoadingScreenState::standard_skirmish(
-                LoadingArtVariant::Americans,
-                0,
-                HouseColorIndex(0),
-                LoadingProgressRowSnapshot {
-                    label: "Player".into(),
-                },
-                cadence,
-            );
-            native.runtime_color_scheme_count = 16;
-            {
-                let mut phase = LoadingPhaseProgress::select(Some(&mut native), true, |_| {
-                    panic!("no atlas must select the real gated sink")
-                });
-                assert!(matches!(&phase.sink, SelectedProgressSink::Gated(_)));
-                assert!(phase.native_theater_cache_mismatch);
-                assert_eq!(phase.runtime_color_scheme_count, 16);
-                for raw in [8, 6, 8, 12, 12] {
-                    phase.sink.milestone(raw);
-                }
-            }
-            assert_eq!(native.progress.current_percent(), expected);
-            {
-                let mut phase = LoadingPhaseProgress::select(Some(&mut native), false, |_| {
-                    panic!("no atlas must select the real gated sink")
-                });
-                assert!(!phase.native_theater_cache_mismatch);
-                phase.sink.milestone(cadence.terminal_raw_percent());
-            }
-            assert_eq!(native.progress.current_percent(), 100.0);
-        }
-    }
-
-    #[test]
-    fn native_loading_state_keeps_the_full_player_progress_ramp() {
-        let mk = |name: &str, hsv: [u8; 3]| ColorSchemeEntry {
-            name: name.into(),
-            hsv,
-        };
-        let ramps = HouseColorRamps::from_schemes(&[
-            mk("Gold", [43, 239, 255]),
-            mk("DarkBlue", [153, 214, 212]),
-        ]);
-        let mut native = NativeLoadingScreenState::standard_skirmish(
-            LoadingArtVariant::Americans,
-            0,
-            HouseColorIndex(1),
-            LoadingProgressRowSnapshot {
-                label: "Player".to_owned(),
-            },
-            NativeLoadingProgressCadence::SelectedMap,
-        );
-        native.resolve_player_colors(
-            &[mk("Gold", [43, 239, 255]), mk("DarkBlue", [153, 214, 212])],
-            &ramps,
-        );
-
-        assert_eq!(native.runtime_color_scheme_count, 4);
-        assert_eq!(native.progress_ramp, *ramps.ramp(HouseColorIndex(1)));
-        assert_ne!(native.progress_ramp[0], native.progress_ramp[15]);
-        assert!(native.progress_ramp[0].b > native.progress_ramp[0].r);
-    }
-
-    #[test]
-    fn progress_row_starts_at_the_first_advancing_milestone() {
-        let mut progress = LoadingProgressState::standard_skirmish();
-        assert_eq!(progress.current_value(), 0.0);
-        assert!(progress.advance_progress(3));
-        assert!(progress.current_value() > 0.0);
-    }
-
-    /// Synthetic `mmpb.shp` frame-0 atlas slot: 12x12 pixels somewhere inside a
-    /// shared atlas, so cropping has to move both the UV origin and the UV size.
-    fn marker_entry() -> LoadingScreenEntry {
-        LoadingScreenEntry {
-            uv_origin: [0.25, 0.5],
-            uv_size: [0.1, 0.2],
-            pixel_size: [12.0, 12.0],
-        }
-    }
-
-    #[test]
-    fn a_preview_wider_than_its_region_is_squashed_whole_not_cropped() {
-        use crate::app::loading::composition::{aspect_fit_preview, mmpb_region_rect};
-
-        // A stock map whose projected preview overruns the 800-wide region: the
-        // fit picks a destination narrower than the source, which is exactly the
-        // case the bar's left-to-right U clamp used to silently crop.
-        let region = mmpb_region_rect(800);
-        let fit = aspect_fit_preview(region, 400, 200).expect("valid fit");
-        assert!(fit.width < 400, "fixture must exercise a downscale");
-
-        let entry = LoadingScreenEntry {
-            uv_origin: [0.5, 0.25],
-            uv_size: [0.4, 0.2],
-            pixel_size: [400.0, 200.0],
-        };
-        let mut instances = Vec::new();
-        push_entry_scaled(
-            &mut instances,
-            entry,
-            [(region.x + fit.pad_x) as f32, (region.y + fit.pad_y) as f32],
-            [fit.width as f32, fit.height as f32],
-            PREVIEW_DEPTH,
-            [1.0; 3],
-        );
-
-        assert_eq!(instances.len(), 1);
-        assert_eq!(instances[0].size, [fit.width as f32, fit.height as f32]);
-        // The whole source is sampled on both axes; nothing is cut off.
-        assert_eq!(instances[0].uv_origin, entry.uv_origin);
-        assert_eq!(instances[0].uv_size, entry.uv_size);
-    }
-
-    #[test]
-    fn the_progress_bar_is_the_only_layer_revealed_by_clipping_u() {
-        let entry = LoadingScreenEntry {
-            uv_origin: [0.0, 0.0],
-            uv_size: [0.4, 0.05],
-            pixel_size: [400.0, 10.0],
-        };
-        let mut instances = Vec::new();
-
-        push_progress_fill(&mut instances, entry, [24.0, 332.0], 100.0, PROGRESS_DEPTH);
-
-        assert_eq!(instances.len(), 1);
-        // A quarter of the frame is uncovered: a quarter of U, all of V.
-        assert_eq!(instances[0].size, [100.0, 10.0]);
-        assert_eq!(instances[0].uv_size, [0.4_f32 * 0.25, 0.05]);
-    }
-
-    #[test]
-    fn both_native_loading_cadences_prepare_scenario_before_first_frame() {
-        assert!(NativeLoadingProgressCadence::SelectedMap.prepares_scenario_before_first_frame());
-        assert!(
-            NativeLoadingProgressCadence::RandomMapHalved.prepares_scenario_before_first_frame()
-        );
-    }
-
-    #[test]
-    fn markers_inside_the_preview_region_draw_uncropped() {
-        let clip = MmpbRegionRect::new(499, 379, 216, 166);
-        let mut instances = Vec::new();
-
-        push_entry_clipped(
-            &mut instances,
-            marker_entry(),
-            [520, 400],
-            MARKER_DEPTH,
-            clip,
-        );
-
-        assert_eq!(instances.len(), 1);
-        assert_eq!(instances[0].position, [520.0, 400.0]);
-        assert_eq!(instances[0].size, [12.0, 12.0]);
-        assert_eq!(instances[0].uv_origin, [0.25, 0.5]);
-        assert_eq!(instances[0].uv_size, [0.1, 0.2]);
-    }
-
-    #[test]
-    fn markers_overhanging_the_preview_region_are_cut_off_at_its_edge() {
-        let clip = MmpbRegionRect::new(499, 379, 216, 166);
-        let mut instances = Vec::new();
-
-        // 4 px past the region's right edge and 3 px above its top edge.
-        push_entry_clipped(
-            &mut instances,
-            marker_entry(),
-            [707, 376],
-            MARKER_DEPTH,
-            clip,
-        );
-
-        assert_eq!(instances.len(), 1);
-        let instance = instances[0];
-        assert_eq!(instance.position, [707.0, 379.0]);
-        assert_eq!(instance.size, [8.0, 9.0]);
-        // The three cropped top rows advance the UV origin; the four cropped
-        // right columns are simply never sampled.
-        assert_eq!(instance.uv_origin, [0.25_f32, 0.5 + 0.2 * 3.0 / 12.0]);
-        assert_eq!(instance.uv_size, [0.1_f32 * 8.0 / 12.0, 0.2 * 9.0 / 12.0]);
-    }
-
-    #[test]
-    fn markers_entirely_outside_the_preview_region_are_dropped() {
-        let clip = MmpbRegionRect::new(499, 379, 216, 166);
-        let mut instances = Vec::new();
-
-        push_entry_clipped(
-            &mut instances,
-            marker_entry(),
-            [715, 400],
-            MARKER_DEPTH,
-            clip,
-        );
-        push_entry_clipped(
-            &mut instances,
-            marker_entry(),
-            [520, 367],
-            MARKER_DEPTH,
-            clip,
-        );
-
-        assert!(instances.is_empty());
-    }
-
-    #[test]
-    fn selected_generic_progress_uses_no_native_loader_metadata() {
-        let mut phase = LoadingPhaseProgress::select(None, true, |_| {
-            panic!("generic loading must never construct rendering progress")
-        });
-        assert!(matches!(&phase.sink, SelectedProgressSink::Generic(_)));
-        assert!(!phase.native_theater_cache_mismatch);
-        assert_eq!(phase.runtime_color_scheme_count, 0);
-        for raw in [8, 6, 8, 12, 100, 200] {
-            phase.sink.milestone(raw);
-        }
-    }
 }
