+++ src/sim/world/command_schedule.rs
@@ -1,0 +1,636 @@
+//! Command admission and scheduled batch execution.
+//!
+//! Preserves envelope insertion order, due filtering, house order and staged
+//! group destinations. The world frame chooses ingress and tail call positions;
+//! `world_commands` implements each individual command payload. Queue state
+//! remains on Simulation for persistence and replay.
+
+use super::Simulation;
+use crate::map::entities::EntityCategory;
+use crate::rules::locomotor_type::LocomotorKind;
+use crate::rules::ruleset::RuleSet;
+use crate::sim::combat;
+use crate::sim::command::{Command, CommandEnvelope};
+use crate::sim::intern::InternedId;
+use crate::sim::movement::locomotor::MovementLayer;
+use crate::sim::movement::{self, group_destination};
+use crate::sim::pathfinding::PathGrid;
+use crate::sim::pathfinding::zone_map::ZoneGrid;
+use std::collections::BTreeMap;
+
+impl Simulation {
+    /// Queue one already-prepared command for future execution.
+    ///
+    /// This is an append-only ingress: it preserves the command owner, execution
+    /// tick, and insertion position exactly. It does not sort, retime, normalize,
+    /// intern, deduplicate, or apply the negotiated input delay.
+    pub fn queue_command(&mut self, command: CommandEnvelope) {
+        self.queue_commands([command]);
+    }
+
+    /// Queue an already-ordered batch for future execution.
+    ///
+    /// Commands are appended in iterator order with their owners, execution
+    /// ticks, and payloads unchanged. Native byte round-trips and network timing
+    /// remain the responsibility of the producer that prepares each envelope.
+    pub fn queue_commands(&mut self, commands: impl IntoIterator<Item = CommandEnvelope>) {
+        self.pending_commands.extend(commands);
+    }
+
+    /// Read-only queue evidence for module tests outside the owning module.
+    #[cfg(test)]
+    pub(crate) fn pending_commands_for_tests(&self) -> &[CommandEnvelope] {
+        &self.pending_commands
+    }
+
+    /// Speed the next ordinary frame will observe after all due offline
+    /// Options transitions execute in canonical house/insertion order.
+    ///
+    /// This read-only projection lets fresh-map and snapshot replacement sync
+    /// client UI without consuming or applying a pending authoritative command.
+    pub(crate) fn projected_in_game_options_speed(&self) -> Option<u8> {
+        let mut projected = u8::try_from(self.session.game_options.game_speed)
+            .ok()
+            .filter(|speed| *speed <= crate::sim::game_options::IN_GAME_OPTIONS_MAX_SPEED);
+        let execute_tick = self.session.tick.saturating_add(1);
+        for owner in self.due_command_house_order(&self.pending_commands, execute_tick) {
+            for command in self
+                .pending_commands
+                .iter()
+                .filter(|command| command.execute_tick <= execute_tick && command.owner == owner)
+            {
+                if !self.houses.contains_key(&owner) {
+                    continue;
+                }
+                let Command::SetGameSpeed { speed } = &command.payload else {
+                    continue;
+                };
+                if *speed <= crate::sim::game_options::IN_GAME_OPTIONS_MAX_SPEED {
+                    projected = Some(*speed);
+                }
+            }
+        }
+        projected
+    }
+
+    /// Consume the terminal edge raised by one executed EXIT command.
+    pub(crate) fn take_executed_exit_owner(&mut self) -> Option<InternedId> {
+        self.executed_exit_owner.take()
+    }
+
+    /// Drain commands that are due for the next tick from `pending_commands`.
+    /// Returns owned commands; remaining commands stay queued.
+    pub fn take_due_commands(&mut self) -> Vec<CommandEnvelope> {
+        let execute_tick = self.session.tick.saturating_add(1);
+        let mut due = Vec::new();
+        let mut kept = Vec::new();
+        for cmd in std::mem::take(&mut self.pending_commands) {
+            if cmd.execute_tick <= execute_tick {
+                due.push(cmd);
+            } else {
+                kept.push(cmd);
+            }
+        }
+        self.pending_commands = kept;
+        due
+    }
+
+    /// Admit and drain one diagnostic-replay batch without consuming pending
+    /// work regenerated by the simulation itself.
+    ///
+    /// The diagnostic log already contains the batch drained by the live app.
+    /// Playback can independently recreate later work such as C4 scatter, so
+    /// mixing that pending work with the recorded batch would admit it twice.
+    /// Temporarily isolating the queue preserves the former direct-replay seam:
+    /// the recorded batch still uses [`Self::queue_commands`] and the canonical
+    /// due filter, malformed future-stamped replay entries are discarded, and
+    /// pre-existing simulation work remains untouched.
+    pub(crate) fn take_due_replay_commands(
+        &mut self,
+        commands: impl IntoIterator<Item = CommandEnvelope>,
+    ) -> Vec<CommandEnvelope> {
+        let simulation_pending = std::mem::take(&mut self.pending_commands);
+        self.queue_commands(commands);
+        let due = self.take_due_commands();
+        self.pending_commands = simulation_pending;
+        due
+    }
+
+    fn apply_one_due_command(
+        &mut self,
+        cmd: &CommandEnvelope,
+        rules: Option<&RuleSet>,
+        path_grid: Option<&PathGrid>,
+        height_map: &BTreeMap<(u16, u16), u8>,
+        overlay_registry: Option<&crate::map::overlay_types::OverlayTypeRegistry>,
+    ) -> (bool, bool, bool, Option<InternedId>) {
+        let cmd_owner_str = self.interner.resolve(cmd.owner).to_string();
+        let applied = self.apply_command_with_overlays(
+            &cmd_owner_str,
+            &cmd.payload,
+            rules,
+            path_grid,
+            height_map,
+            overlay_registry,
+        );
+        let placed_building_owner = self.successful_non_wall_placement_owner(cmd, applied, rules);
+        let spawned_entity = placed_building_owner.is_some()
+            || applied
+                && matches!(
+                    cmd.payload,
+                    Command::DeployMcv { .. }
+                        | Command::UndeployBuilding { .. }
+                        | Command::LaunchSuperWeapon { .. }
+                );
+        let destroyed_structure = applied
+            && matches!(
+                cmd.payload,
+                Command::SellBuilding { .. } | Command::UndeployBuilding { .. }
+            );
+        (
+            applied,
+            spawned_entity,
+            destroyed_structure,
+            placed_building_owner,
+        )
+    }
+
+    pub(super) fn successful_non_wall_placement_owner(
+        &self,
+        command: &CommandEnvelope,
+        applied: bool,
+        rules: Option<&RuleSet>,
+    ) -> Option<InternedId> {
+        let Command::PlaceReadyBuilding { type_id, .. } = command.payload else {
+            return None;
+        };
+        (applied
+            && rules
+                .and_then(|rules| rules.object(self.interner.resolve(type_id)))
+                .is_some_and(|object| !object.wall))
+        .then_some(command.owner)
+    }
+
+    pub(super) fn is_wall_placement_command(
+        &self,
+        command: &Command,
+        rules: Option<&RuleSet>,
+    ) -> bool {
+        let Command::PlaceReadyBuilding { type_id, .. } = command else {
+            return false;
+        };
+        rules
+            .and_then(|rules| rules.object(self.interner.resolve(*type_id)))
+            .is_some_and(|object| object.wall)
+    }
+
+    fn command_uses_megamission(command: &Command) -> bool {
+        matches!(
+            command,
+            Command::Move { .. }
+                | Command::Attack { .. }
+                | Command::ForceAttack { .. }
+                | Command::AttackMove { .. }
+                | Command::Guard { .. }
+                | Command::MinerReturn { .. }
+                | Command::RepairAtDepot { .. }
+                | Command::EnterTransport { .. }
+                | Command::EnterBunker { .. }
+                | Command::EjectBunker { .. }
+                | Command::UnloadPassengers { .. }
+                | Command::HarvestCell { .. }
+                | Command::CaptureBuilding { .. }
+                | Command::PlantC4 { .. }
+                | Command::ToggleInfantryDeploy { .. }
+                | Command::ForceAttackCell { .. }
+        )
+    }
+
+    fn formation_key(command: &Command) -> Option<(u8, u16, u16)> {
+        match command {
+            Command::Move {
+                target_rx,
+                target_ry,
+                ..
+            } => Some((0, *target_rx, *target_ry)),
+            Command::AttackMove {
+                target_rx,
+                target_ry,
+                ..
+            } => Some((1, *target_rx, *target_ry)),
+            _ => None,
+        }
+    }
+
+    fn formation_entity_id(command: &Command) -> Option<u64> {
+        match command {
+            Command::Move { entity_id, .. } | Command::AttackMove { entity_id, .. } => {
+                Some(*entity_id)
+            }
+            _ => None,
+        }
+    }
+
+    fn group_destination_candidate_facts(
+        &self,
+        grid: &PathGrid,
+        zone_grid: &ZoneGrid,
+        clicked_target: (i16, i16),
+        member: &group_destination::GroupDestinationMember,
+        candidate: (i16, i16),
+    ) -> group_destination::CandidateFacts {
+        let signed_candidate = (i32::from(candidate.0), i32::from(candidate.1));
+        if !crate::sim::cell_rect::cell_is_in_playfield_height_aware(
+            signed_candidate,
+            self.playfield_bounds,
+            self.resolved_terrain.as_ref(),
+        ) {
+            return group_destination::CandidateFacts::outside_playfield();
+        }
+        let (Ok(candidate_x), Ok(candidate_y)) =
+            (u16::try_from(candidate.0), u16::try_from(candidate.1))
+        else {
+            return group_destination::CandidateFacts::outside_playfield();
+        };
+        let target = (clicked_target.0 as u16, clicked_target.1 as u16);
+        let Some(entity) = self.substrate.entities.get(member.entity_id) else {
+            return group_destination::CandidateFacts {
+                in_playfield: true,
+                same_zone: false,
+                height_band_ok: false,
+                can_enter_code: 7,
+            };
+        };
+        let Some(target_cell) = grid.cell(target.0, target.1) else {
+            return group_destination::CandidateFacts {
+                in_playfield: true,
+                same_zone: false,
+                height_band_ok: false,
+                can_enter_code: 7,
+            };
+        };
+        let Some(candidate_cell) = grid.cell(candidate_x, candidate_y) else {
+            return group_destination::CandidateFacts {
+                in_playfield: true,
+                same_zone: false,
+                height_band_ok: false,
+                can_enter_code: 7,
+            };
+        };
+
+        let movement_zone = entity
+            .locomotor
+            .as_ref()
+            .map(|locomotor| locomotor.movement_zone)
+            .unwrap_or_default();
+        let speed_type = entity
+            .locomotor
+            .as_ref()
+            .map(|locomotor| locomotor.speed_type)
+            .unwrap_or_default();
+        let locomotor_kind = entity
+            .locomotor
+            .as_ref()
+            .map_or(LocomotorKind::Drive, |locomotor| locomotor.effective_kind());
+        let target_layer = if target_cell.has_structural_bridge() {
+            MovementLayer::Bridge
+        } else {
+            MovementLayer::Ground
+        };
+        let candidate_layer = if candidate_cell.has_structural_bridge() {
+            MovementLayer::Bridge
+        } else {
+            MovementLayer::Ground
+        };
+        let same_zone = zone_grid.map_for(movement_zone).is_some_and(|zones| {
+            zones.zone_at(target.0, target.1, target_layer)
+                == zones.zone_at(candidate_x, candidate_y, candidate_layer)
+        });
+
+        let target_height = target_cell.signed_level()
+            + if target_cell.has_structural_bridge() {
+                4
+            } else {
+                0
+            };
+        let candidate_height = candidate_cell.signed_level()
+            + if candidate_cell.has_structural_bridge() {
+                4
+            } else {
+                0
+            };
+        let height_band_ok = (candidate_height - target_height).abs() < 3;
+
+        // UnitClass::Can_Enter_Cell(candidate, -1, target_height, 0, 1)
+        // selects the bridge list when the candidate carries a structural deck
+        // separated from the requested path height.
+        let object_layer = if candidate_cell.has_structural_bridge()
+            && (target_height - candidate_cell.signed_level()).abs() >= 2
+        {
+            MovementLayer::Bridge
+        } else {
+            MovementLayer::Ground
+        };
+        let layers = crate::sim::pathfinding::can_enter_layer_context(
+            object_layer,
+            object_layer,
+            candidate_cell,
+            target_height,
+        );
+        let cost_grid = self.terrain_costs.get(&speed_type);
+        let terrain_passable = match layers.terrain_layer {
+            MovementLayer::Ground => {
+                crate::sim::pathfinding::is_cell_passable_for_mover_with_speed(
+                    grid,
+                    candidate_x,
+                    candidate_y,
+                    Some(movement_zone),
+                    Some(speed_type),
+                    self.resolved_terrain.as_ref(),
+                    cost_grid,
+                    false,
+                    crate::sim::pathfinding::cell_entry::TerrainEntryMode::Smoothing,
+                )
+            }
+            MovementLayer::Bridge => {
+                grid.is_walkable_on_layer(candidate_x, candidate_y, MovementLayer::Bridge)
+            }
+            MovementLayer::Air | MovementLayer::Underground => true,
+        };
+        let can_enter_code = if !terrain_passable {
+            7
+        } else {
+            match crate::sim::pathfinding::cell_entry::check_terrain_with_layers(
+                (candidate_x, candidate_y),
+                layers,
+                entity.category,
+                Some(grid),
+                cost_grid,
+                &self.substrate.occupancy,
+            ) {
+                crate::sim::pathfinding::cell_entry::TerrainCheckResult::Clear => 0,
+                crate::sim::pathfinding::cell_entry::TerrainCheckResult::Impassable => 7,
+                crate::sim::pathfinding::cell_entry::TerrainCheckResult::NeedsBlockerCheck => {
+                    crate::sim::pathfinding::cell_entry::classify_occupied_cell_with_layers(
+                        (candidate_x, candidate_y),
+                        layers,
+                        member.entity_id,
+                        movement::bump_crush::CrushCapability::new(
+                            entity.regular_crusher,
+                            entity.omni_crusher,
+                        ),
+                        self.interner.resolve(entity.owner()),
+                        locomotor_kind,
+                        false,
+                        &self.substrate.occupancy,
+                        &self.substrate.entities,
+                        &self.house_alliances,
+                        &self.interner,
+                    )
+                    .yr_code()
+                }
+            }
+        };
+
+        group_destination::CandidateFacts {
+            in_playfield: true,
+            same_zone,
+            height_band_ok,
+            can_enter_code,
+        }
+    }
+
+    /// Adjust consecutive same-target movement runs after their house's
+    /// non-megamission scan, immediately before staged command execution.
+    pub(super) fn adjust_staged_megamission_destinations(
+        &self,
+        commands: &mut [CommandEnvelope],
+        path_grid: Option<&PathGrid>,
+    ) {
+        let (Some(grid), Some(zone_grid)) = (path_grid, self.zone_grid.as_ref()) else {
+            return;
+        };
+        let mut run_start = 0;
+        while run_start < commands.len() {
+            let Some(key) = Self::formation_key(&commands[run_start].payload) else {
+                run_start += 1;
+                continue;
+            };
+            let mut run_end = run_start + 1;
+            while run_end < commands.len()
+                && Self::formation_key(&commands[run_end].payload) == Some(key)
+            {
+                run_end += 1;
+            }
+            if run_end - run_start > 1 {
+                let mut members = Vec::new();
+                for (command_index, command) in commands[run_start..run_end].iter().enumerate() {
+                    let Some(entity_id) = Self::formation_entity_id(&command.payload) else {
+                        continue;
+                    };
+                    let Some(entity) = self.substrate.entities.get(entity_id) else {
+                        continue;
+                    };
+                    if !matches!(
+                        entity.category,
+                        EntityCategory::Unit | EntityCategory::Infantry | EntityCategory::Aircraft
+                    ) || entity.low_bridge_tube_state.is_some()
+                    {
+                        continue;
+                    }
+                    let Some(coord_z) = self
+                        .resolved_terrain
+                        .as_ref()
+                        .and_then(|terrain| combat::in_range::effective_z_leptons(entity, terrain))
+                    else {
+                        continue;
+                    };
+                    members.push(group_destination::GroupDestinationMember {
+                        command_index,
+                        entity_id,
+                        coord: [
+                            i32::from(entity.position.rx)
+                                .wrapping_mul(256)
+                                .wrapping_add(entity.position.sub_x.to_num::<i32>()),
+                            i32::from(entity.position.ry)
+                                .wrapping_mul(256)
+                                .wrapping_add(entity.position.sub_y.to_num::<i32>()),
+                            coord_z as i32,
+                        ],
+                        source_cell: (entity.position.rx as i16, entity.position.ry as i16),
+                    });
+                }
+                let assignments = group_destination::distribute_group_destinations(
+                    (key.1 as i16, key.2 as i16),
+                    &members,
+                    |member, candidate| {
+                        self.group_destination_candidate_facts(
+                            grid,
+                            zone_grid,
+                            (key.1 as i16, key.2 as i16),
+                            member,
+                            candidate,
+                        )
+                    },
+                );
+                for assignment in assignments {
+                    let Some(command) = commands.get_mut(run_start + assignment.command_index)
+                    else {
+                        continue;
+                    };
+                    match &mut command.payload {
+                        Command::Move {
+                            target_rx,
+                            target_ry,
+                            ..
+                        }
+                        | Command::AttackMove {
+                            target_rx,
+                            target_ry,
+                            ..
+                        } => {
+                            *target_rx = assignment.destination.0 as u16;
+                            *target_ry = assignment.destination.1 as u16;
+                        }
+                        _ => {}
+                    }
+                }
+            }
+            run_start = run_end;
+        }
+    }
+
+    /// Canonical receiver order for every due command in one master frame.
+    /// Registered HouseClass order wins; standalone-fixture owners are appended
+    /// in their first issue order, matching the established tail dispatcher.
+    fn due_command_house_order(
+        &self,
+        commands: &[CommandEnvelope],
+        execute_tick: u64,
+    ) -> Vec<InternedId> {
+        let mut house_order = self.session.house_order.clone();
+        for command in commands
+            .iter()
+            .filter(|command| command.execute_tick <= execute_tick)
+        {
+            if !house_order.contains(&command.owner) {
+                house_order.push(command.owner);
+            }
+        }
+        house_order
+    }
+
+    fn command_uses_frame_ingress(command: &Command) -> bool {
+        matches!(command, Command::SetGameSpeed { .. })
+    }
+
+    /// Apply offline session transitions before triggers and the live-object
+    /// walk. Native offline Options stores GameSpeed before the next Main_Tick;
+    /// VERA transports the transition as a replayable command, then admits it
+    /// at this dedicated ingress instead of the ordinary EventClass tail.
+    pub(super) fn apply_due_frame_ingress_commands(
+        &mut self,
+        commands: &[CommandEnvelope],
+        execute_tick: u64,
+    ) -> usize {
+        let mut executed_commands = 0usize;
+        for owner in self.due_command_house_order(commands, execute_tick) {
+            for command in commands.iter().filter(|command| {
+                command.execute_tick <= execute_tick
+                    && command.owner == owner
+                    && Self::command_uses_frame_ingress(&command.payload)
+            }) {
+                let Command::SetGameSpeed { speed } = &command.payload else {
+                    unreachable!("frame-ingress predicate admitted a non-session command");
+                };
+                if self.houses.contains_key(&owner) {
+                    let _ = self.session.game_options.apply_in_game_speed(*speed);
+                }
+                // Preserve the established dispatcher convention: every due
+                // envelope is consumed/counts even when validation rejects it.
+                executed_commands += 1;
+            }
+        }
+        executed_commands
+    }
+
+    /// Apply all due tail commands in HouseClass registration
+    /// order. Each house preserves insertion order within the normal and
+    /// staged-megamission streams. Returns
+    /// `(executed_commands, spawned_entities, destroyed_structure,
+    /// successful_non_wall_placement_owners)`.
+    pub(super) fn apply_due_commands(
+        &mut self,
+        commands: &[CommandEnvelope],
+        rules: Option<&RuleSet>,
+        path_grid: Option<&PathGrid>,
+        height_map: &BTreeMap<(u16, u16), u8>,
+        execute_tick: u64,
+        overlay_registry: Option<&crate::map::overlay_types::OverlayTypeRegistry>,
+    ) -> (usize, bool, bool, Vec<InternedId>) {
+        let mut executed_commands = 0usize;
+        let mut spawned_entities = false;
+        let mut destroyed_structure = false;
+        let mut placed_building_owners = Vec::new();
+        let mut tail_path_grid = path_grid
+            .cloned()
+            .or_else(|| self.path_grid.as_deref().cloned());
+
+        for owner in self.due_command_house_order(commands, execute_tick) {
+            for command in commands.iter().filter(|command| {
+                command.execute_tick <= execute_tick
+                    && command.owner == owner
+                    && !Self::command_uses_frame_ingress(&command.payload)
+                    && !Self::command_uses_megamission(&command.payload)
+            }) {
+                let (applied, spawned, destroyed, placed_owner) = self.apply_one_due_command(
+                    command,
+                    rules,
+                    tail_path_grid.as_ref(),
+                    height_map,
+                    overlay_registry,
+                );
+                if matches!(command.payload, Command::SellWallAtCell { .. }) {
+                    tail_path_grid = self.path_grid.as_deref().cloned();
+                } else if applied && self.is_wall_placement_command(&command.payload, rules) {
+                    tail_path_grid = self.path_grid.as_deref().cloned().or(tail_path_grid);
+                }
+                spawned_entities |= spawned;
+                destroyed_structure |= destroyed;
+                placed_building_owners.extend(placed_owner);
+                executed_commands += 1;
+            }
+
+            let mut staged = commands
+                .iter()
+                .filter(|command| {
+                    command.execute_tick <= execute_tick
+                        && command.owner == owner
+                        && Self::command_uses_megamission(&command.payload)
+                })
+                .cloned()
+                .collect::<Vec<_>>();
+            self.adjust_staged_megamission_destinations(&mut staged, tail_path_grid.as_ref());
+            for command in &staged {
+                let (_, spawned, destroyed, placed_owner) = self.apply_one_due_command(
+                    command,
+                    rules,
+                    tail_path_grid.as_ref(),
+                    height_map,
+                    overlay_registry,
+                );
+                spawned_entities |= spawned;
+                destroyed_structure |= destroyed;
+                placed_building_owners.extend(placed_owner);
+                executed_commands += 1;
+            }
+        }
+
+        (
+            executed_commands,
+            spawned_entities,
+            destroyed_structure,
+            placed_building_owners,
+        )
+    }
+}
